refactor(cli): extract monitoring, system, doctor, and maintenance groups
Move the remaining command handlers out of main.rs:
- monitoring: security + audit + memory + devices + webhooks
- system: system info/version, scaffold, completion, dashboard, migrate
- doctor_cmd: the `doctor` diagnostic command (kept separate from the
  pre-existing `doctor` helper module)
- maintenance: update + uninstall + reset + service, with the update
  endpoint constants and the UpdateLaunch / ReleaseComparison enums.

main.rs is now ~1.8k lines (from ~14.6k): main(), tracing/process
setup, and the dispatch match. Pure code move, no behavior change.
