+++ crates/librefang-cli/src/main.rs
@@ -670,3136 +670,6 @@ fn main() {
     }
 }
 
-fn cmd_doctor(json: bool, repair: bool) {
-    // BrokenPipe protection for the WHOLE command, not just the --json
-    // branch. `librefang doctor | head -5` and similar pipelines drop the
-    // reader after a few lines, which on the next stdout write turns into a
-    // panic — Rust ignores SIGPIPE by default and translates EPIPE into an
-    // io::Error that `println!` unwraps.
-    //
-    // The pre-existing `write_stdout_safe` helper only covered the
-    // `--json` final emission. Hundreds of `ui::*` and bare `println!`
-    // calls between the start of cmd_doctor and that emission were still
-    // unprotected. Restoring the default SIGPIPE handler for the duration
-    // of this command makes the kernel terminate the process cleanly on
-    // pipe close instead, covering every print path in this function and
-    // the `ui::*` helpers it calls.
-    #[cfg(unix)]
-    unsafe {
-        libc::signal(libc::SIGPIPE, libc::SIG_DFL);
-    }
-
-    let mut checks: Vec<serde_json::Value> = Vec::new();
-    let mut all_ok = true;
-    let mut repaired = false;
-
-    if !json {
-        ui::step(&i18n::t("doctor-title"));
-        println!();
-    }
-
-    let home = dirs::home_dir();
-    if let Some(_h) = &home {
-        let librefang_dir = cli_librefang_home();
-
-        // --- Check 1: LibreFang directory ---
-        if librefang_dir.exists() {
-            if !json {
-                ui::check_ok(&format!("LibreFang directory: {}", librefang_dir.display()));
-            }
-            checks.push(serde_json::json!({"check": "librefang_dir", "status": "ok", "path": librefang_dir.display().to_string()}));
-        } else if repair {
-            if !json {
-                ui::check_fail("LibreFang directory not found.");
-            }
-            let answer = prompt_input("    Create it now? [Y/n] ");
-            if answer.is_empty() || answer.starts_with('y') || answer.starts_with('Y') {
-                if std::fs::create_dir_all(&librefang_dir).is_ok() {
-                    restrict_dir_permissions(&librefang_dir);
-                    let _ = std::fs::create_dir_all(librefang_dir.join("data"));
-                    let _ =
-                        std::fs::create_dir_all(librefang_dir.join("workspaces").join("agents"));
-                    if !json {
-                        ui::check_ok("Created LibreFang directory");
-                    }
-                    repaired = true;
-                } else {
-                    if !json {
-                        ui::check_fail("Failed to create directory");
-                    }
-                    all_ok = false;
-                }
-            } else {
-                all_ok = false;
-            }
-            checks.push(serde_json::json!({"check": "librefang_dir", "status": if repaired { "repaired" } else { "fail" }}));
-        } else {
-            if !json {
-                ui::check_fail("LibreFang directory not found. Run `librefang init` first.");
-            }
-            checks.push(serde_json::json!({"check": "librefang_dir", "status": "fail"}));
-            all_ok = false;
-        }
-
-        // --- Check 2: .env file exists + permissions ---
-        let env_path = librefang_dir.join(".env");
-        if env_path.exists() {
-            #[cfg(unix)]
-            {
-                use std::os::unix::fs::PermissionsExt;
-                if let Ok(meta) = std::fs::metadata(&env_path) {
-                    let mode = meta.permissions().mode() & 0o777;
-                    if mode == 0o600 {
-                        if !json {
-                            ui::check_ok(".env file (permissions OK)");
-                        }
-                    } else if repair {
-                        let _ = std::fs::set_permissions(
-                            &env_path,
-                            std::fs::Permissions::from_mode(0o600),
-                        );
-                        if !json {
-                            ui::check_ok(".env file (permissions fixed to 0600)");
-                        }
-                        repaired = true;
-                    } else if !json {
-                        ui::check_warn(&format!(
-                            ".env file has loose permissions ({:o}), should be 0600",
-                            mode
-                        ));
-                    }
-                } else if !json {
-                    ui::check_ok(".env file");
-                }
-            }
-            #[cfg(not(unix))]
-            {
-                if !json {
-                    ui::check_ok(".env file");
-                }
-            }
-            checks.push(serde_json::json!({"check": "env_file", "status": "ok"}));
-        } else {
-            if !json {
-                ui::check_warn(
-                    ".env file not found (create with: librefang config set-key <provider>)",
-                );
-            }
-            checks.push(serde_json::json!({"check": "env_file", "status": "warn"}));
-        }
-
-        // --- Check 3: Config TOML syntax validation ---
-        let config_path = librefang_dir.join("config.toml");
-        if config_path.exists() {
-            let config_content = std::fs::read_to_string(&config_path).unwrap_or_default();
-            match toml::from_str::<toml::Value>(&config_content) {
-                Ok(_) => {
-                    if !json {
-                        ui::check_ok(&format!("Config file: {}", config_path.display()));
-                    }
-                    checks.push(serde_json::json!({"check": "config_file", "status": "ok"}));
-                }
-                Err(e) => {
-                    if !json {
-                        ui::check_fail(&format!("Config file has syntax errors: {e}"));
-                        ui::hint(&i18n::t("hint-config-edit"));
-                    }
-                    checks.push(serde_json::json!({"check": "config_syntax", "status": "fail", "error": e.to_string()}));
-                    all_ok = false;
-                }
-            }
-        } else if repair {
-            if !json {
-                ui::check_fail("Config file not found.");
-            }
-            let answer = prompt_input("    Create default config? [Y/n] ");
-            if answer.is_empty() || answer.starts_with('y') || answer.starts_with('Y') {
-                let (provider, api_key_env, model) = detect_best_provider();
-                let default_config = render_init_default_config(&provider, &model, &api_key_env);
-                let _ = std::fs::create_dir_all(&librefang_dir);
-                if std::fs::write(&config_path, default_config).is_ok() {
-                    restrict_file_permissions(&config_path);
-                    if !json {
-                        ui::check_ok("Created default config.toml");
-                    }
-                    repaired = true;
-                } else {
-                    if !json {
-                        ui::check_fail("Failed to create config.toml");
-                    }
-                    all_ok = false;
-                }
-            } else {
-                all_ok = false;
-            }
-            checks.push(serde_json::json!({"check": "config_file", "status": if repaired { "repaired" } else { "fail" }}));
-        } else {
-            if !json {
-                ui::check_fail("Config file not found.");
-            }
-            checks.push(serde_json::json!({"check": "config_file", "status": "fail"}));
-            all_ok = false;
-        }
-
-        // --- Check: Version update ---
-        {
-            let current_version = env!("CARGO_PKG_VERSION");
-            let update_channel = load_update_channel_from_config().unwrap_or_default();
-            if !json {
-                ui::check_ok(&format!(
-                    "CLI version: {current_version} (channel: {update_channel})"
-                ));
-            }
-            checks.push(serde_json::json!({"check": "cli_version", "status": "ok", "version": current_version, "channel": update_channel.to_string()}));
-
-            // Try to fetch latest release for the configured channel (best-effort)
-            match fetch_latest_release_tag(update_channel) {
-                Ok(tag) => {
-                    let latest = tag.strip_prefix('v').unwrap_or(&tag);
-                    if latest != current_version {
-                        if !json {
-                            ui::check_warn(&format!(
-                                "Update available: {current_version} -> {latest} (see https://github.com/librefang/librefang/releases)"
-                            ));
-                        }
-                        checks.push(serde_json::json!({"check": "version_update", "status": "warn", "current": current_version, "latest": latest}));
-                    } else {
-                        if !json {
-                            ui::check_ok("CLI is up to date");
-                        }
-                        checks.push(serde_json::json!({"check": "version_update", "status": "ok"}));
-                    }
-                }
-                Err(_) => {
-                    if !json {
-                        ui::check_warn("Could not check for updates (network unavailable)");
-                    }
-                    checks.push(serde_json::json!({"check": "version_update", "status": "warn", "reason": "network_error"}));
-                }
-            }
-        }
-
-        // --- Check 4: Port availability ---
-        // Read api_listen from config (default: 127.0.0.1:4545)
-        let api_listen = {
-            let cfg_path = librefang_dir.join("config.toml");
-            if cfg_path.exists() {
-                std::fs::read_to_string(&cfg_path)
-                    .ok()
-                    .and_then(|s| toml::from_str::<librefang_types::config::KernelConfig>(&s).ok())
-                    .map(|c| c.api_listen)
-                    .unwrap_or_else(|| librefang_types::config::DEFAULT_API_LISTEN.to_string())
-            } else {
-                librefang_types::config::DEFAULT_API_LISTEN.to_string()
-            }
-        };
-        if !json {
-            println!();
-        }
-        let daemon_running = find_daemon();
-        if let Some(ref base) = daemon_running {
-            if !json {
-                ui::check_ok(&format!("Daemon running at {base}"));
-            }
-            checks.push(serde_json::json!({"check": "daemon", "status": "ok", "url": base}));
-        } else {
-            if !json {
-                ui::check_warn("Daemon not running (start with `librefang start`)");
-            }
-            checks.push(serde_json::json!({"check": "daemon", "status": "warn"}));
-
-            // Check if the configured port is available
-            let bind_addr = if api_listen.starts_with("0.0.0.0") {
-                api_listen.replacen("0.0.0.0", "127.0.0.1", 1)
-            } else {
-                api_listen.clone()
-            };
-            match std::net::TcpListener::bind(&bind_addr) {
-                Ok(_) => {
-                    if !json {
-                        ui::check_ok(&format!("Port {api_listen} is available"));
-                    }
-                    checks.push(
-                        serde_json::json!({"check": "port", "status": "ok", "address": api_listen}),
-                    );
-                }
-                Err(_) => {
-                    if !json {
-                        ui::check_warn(&format!("Port {api_listen} is in use by another process"));
-                    }
-                    checks.push(serde_json::json!({"check": "port", "status": "warn", "address": api_listen}));
-                }
-            }
-        }
-
-        // --- Check 5: Stale daemon.json ---
-        let daemon_json_path = librefang_dir.join("daemon.json");
-        if daemon_json_path.exists() && daemon_running.is_none() {
-            if repair {
-                let _ = std::fs::remove_file(&daemon_json_path);
-                if !json {
-                    ui::check_ok("Removed stale daemon.json");
-                }
-                repaired = true;
-            } else if !json {
-                ui::check_warn(
-                    "Stale daemon.json found (daemon not running). Run with --repair to clean up.",
-                );
-            }
-            checks.push(serde_json::json!({"check": "stale_daemon_json", "status": if repair { "repaired" } else { "warn" }}));
-        }
-
-        // --- Check 6: Database file ---
-        let db_path = librefang_dir.join("data").join("librefang.db");
-        if db_path.exists() {
-            // Quick SQLite magic bytes check
-            if let Ok(bytes) = std::fs::read(&db_path) {
-                if bytes.len() >= 16 && bytes.starts_with(b"SQLite format 3") {
-                    if !json {
-                        ui::check_ok("Database file (valid SQLite)");
-                    }
-                    checks.push(serde_json::json!({"check": "database", "status": "ok"}));
-                } else {
-                    if !json {
-                        ui::check_fail("Database file exists but is not valid SQLite");
-                    }
-                    checks.push(serde_json::json!({"check": "database", "status": "fail"}));
-                    all_ok = false;
-                }
-            }
-        } else {
-            if !json {
-                ui::check_warn("No database file (will be created on first run)");
-            }
-            checks.push(serde_json::json!({"check": "database", "status": "warn"}));
-        }
-
-        // --- Check 7: Disk space ---
-        #[cfg(unix)]
-        {
-            if let Ok(output) = std::process::Command::new("df")
-                .args(["-m", &librefang_dir.display().to_string()])
-                .output()
-            {
-                let stdout = String::from_utf8_lossy(&output.stdout);
-                // Parse the available MB from df output (4th column of 2nd line)
-                if let Some(line) = stdout.lines().nth(1) {
-                    let cols: Vec<&str> = line.split_whitespace().collect();
-                    if cols.len() >= 4 {
-                        if let Ok(available_mb) = cols[3].parse::<u64>() {
-                            if available_mb < 100 {
-                                if !json {
-                                    ui::check_warn(&format!(
-                                        "Low disk space: {available_mb}MB available"
-                                    ));
-                                }
-                                checks.push(serde_json::json!({"check": "disk_space", "status": "warn", "available_mb": available_mb}));
-                            } else {
-                                if !json {
-                                    ui::check_ok(&format!(
-                                        "Disk space: {available_mb}MB available"
-                                    ));
-                                }
-                                checks.push(serde_json::json!({"check": "disk_space", "status": "ok", "available_mb": available_mb}));
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // --- Check 8: Agent manifests parse correctly ---
-        let agents_dir = librefang_dir.join("workspaces").join("agents");
-        if agents_dir.exists() {
-            let mut agent_errors = Vec::new();
-            if let Ok(entries) = std::fs::read_dir(&agents_dir) {
-                for entry in entries.flatten() {
-                    let path = entry.path();
-                    if path.extension().and_then(|e| e.to_str()) == Some("toml") {
-                        if let Ok(content) = std::fs::read_to_string(&path) {
-                            if let Err(e) = toml::from_str::<AgentManifest>(&content) {
-                                agent_errors.push((
-                                    path.file_name()
-                                        .unwrap_or_default()
-                                        .to_string_lossy()
-                                        .to_string(),
-                                    e.to_string(),
-                                ));
-                            }
-                        }
-                    }
-                }
-            }
-            if agent_errors.is_empty() {
-                if !json {
-                    ui::check_ok("Agent manifests are valid");
-                }
-                checks.push(serde_json::json!({"check": "agent_manifests", "status": "ok"}));
-            } else {
-                for (file, err) in &agent_errors {
-                    if !json {
-                        ui::check_fail(&format!("Invalid manifest {file}: {err}"));
-                    }
-                }
-                checks.push(serde_json::json!({"check": "agent_manifests", "status": "fail", "errors": agent_errors.len()}));
-                all_ok = false;
-            }
-        }
-    } else {
-        if !json {
-            ui::check_fail("Could not determine home directory");
-        }
-        checks.push(serde_json::json!({"check": "home_dir", "status": "fail"}));
-        all_ok = false;
-    }
-
-    // --- LLM providers ---
-    if !json {
-        println!("\n  LLM Providers:");
-    }
-    // Pretty display names for known provider IDs. Anything not listed
-    // here falls back to a Title-Case derivation of the raw provider id
-    // (e.g. `xiaomi` → `Xiaomi`). Adding a new provider to
-    // `PROVIDER_REGISTRY` automatically picks up the fallback so the
-    // check loop never silently misses a key — only the cosmetic name
-    // needs editing here, not the list of providers checked.
-    fn display_name(provider_id: &str) -> String {
-        match provider_id {
-            "openai" => "OpenAI".to_string(),
-            "openrouter" => "OpenRouter".to_string(),
-            "deepseek" => "DeepSeek".to_string(),
-            "deepinfra" => "DeepInfra".to_string(),
-            "byteplus" => "BytePlus".to_string(),
-            "azure-openai" => "Azure OpenAI".to_string(),
-            "github-copilot" => "GitHub Copilot".to_string(),
-            "huggingface" => "Hugging Face".to_string(),
-            "openai-codex" => "OpenAI Codex".to_string(),
-            "claude-code" => "Claude Code".to_string(),
-            "vertex-ai" => "Vertex AI".to_string(),
-            "nvidia-nim" => "NVIDIA NIM".to_string(),
-            "z.ai" | "zai" => "Z.ai".to_string(),
-            "kimi-coding" | "kimi_coding" => "Kimi Coding".to_string(),
-            "alibaba-coding-plan" => "Alibaba Coding Plan".to_string(),
-            other => {
-                // Title-case fallback for unlisted providers so `xiaomi` →
-                // `Xiaomi` instead of leaking the raw lowercase id.
-                let mut chars = other.chars();
-                match chars.next() {
-                    Some(c) => c.to_uppercase().collect::<String>() + chars.as_str(),
-                    None => String::new(),
-                }
-            }
-        }
-    }
-
-    // Drive doctor off PROVIDER_REGISTRY so adding a provider to the
-    // driver layer never requires a parallel edit here. `GOOGLE_API_KEY`
-    // (gemini's alt env) and similar aliases come through automatically.
-    // This subsumes the previous hardcoded array (including the byteplus
-    // entry from #3274 — now provided automatically by the registry).
-    let provider_specs = librefang_runtime::drivers::cloud_provider_key_specs();
-    let provider_keys: Vec<(&str, String, &str)> = provider_specs
-        .iter()
-        .map(|(env_var, provider_id)| (*env_var, display_name(provider_id), *provider_id))
-        .collect();
-
-    let mut any_key_set = false;
-    for (env_var, name, provider_id) in &provider_keys {
-        let set = std::env::var(env_var).is_ok();
-        if set {
-            // --- Check 9: Live key validation ---
-            let valid = test_api_key(provider_id, &std::env::var(env_var).unwrap_or_default());
-            if valid {
-                if !json {
-                    ui::provider_status(name, env_var, true);
-                }
-            } else if !json {
-                ui::check_warn(&format!("{name} ({env_var}) - key rejected (401/403)"));
-            }
-            any_key_set = true;
-            checks.push(serde_json::json!({"check": "provider", "name": name, "env_var": env_var, "status": if valid { "ok" } else { "warn" }, "live_test": !valid}));
-        } else {
-            if !json {
-                ui::provider_status(name, env_var, false);
-            }
-            checks.push(serde_json::json!({"check": "provider", "name": name, "env_var": env_var, "status": "warn"}));
-        }
-    }
-
-    if !any_key_set {
-        if !json {
-            println!();
-            ui::check_fail(&i18n::t("doctor-no-api-keys"));
-            ui::blank();
-            ui::section(&i18n::t("section-getting-api-key"));
-            ui::suggest_cmd("Groq:", "https://console.groq.com       (free, fast)");
-            ui::suggest_cmd("Gemini:", "https://aistudio.google.com    (free tier)");
-            ui::suggest_cmd("DeepSeek:", "https://platform.deepseek.com  (low cost)");
-            ui::blank();
-            ui::hint(&i18n::t("hint-set-key"));
-        }
-        all_ok = false;
-    }
-
-    // --- Check: Network connectivity to configured LLM provider endpoints ---
-    {
-        let provider_endpoints: &[(&str, &str, &str)] = &[
-            ("OPENAI_API_KEY", "OpenAI", "api.openai.com:443"),
-            ("ANTHROPIC_API_KEY", "Anthropic", "api.anthropic.com:443"),
-            ("GROQ_API_KEY", "Groq", "api.groq.com:443"),
-            ("DEEPSEEK_API_KEY", "DeepSeek", "api.deepseek.com:443"),
-            (
-                "GEMINI_API_KEY",
-                "Gemini",
-                "generativelanguage.googleapis.com:443",
-            ),
-            (
-                "GOOGLE_API_KEY",
-                "Google",
-                "generativelanguage.googleapis.com:443",
-            ),
-            ("OPENROUTER_API_KEY", "OpenRouter", "openrouter.ai:443"),
-            ("TOGETHER_API_KEY", "Together", "api.together.xyz:443"),
-            ("MISTRAL_API_KEY", "Mistral", "api.mistral.ai:443"),
-            ("FIREWORKS_API_KEY", "Fireworks", "api.fireworks.ai:443"),
-        ];
-
-        let configured: Vec<_> = provider_endpoints
-            .iter()
-            .filter(|(env_var, _, _)| std::env::var(env_var).is_ok())
-            .collect();
-
-        if !configured.is_empty() {
-            if !json {
-                println!("\n  Network Connectivity:");
-            }
-            for (env_var, name, endpoint) in &configured {
-                use std::net::{TcpStream, ToSocketAddrs};
-                let reachable = endpoint
-                    .to_socket_addrs()
-                    .ok()
-                    .and_then(|mut addrs| addrs.next())
-                    .map(|addr| {
-                        TcpStream::connect_timeout(&addr, std::time::Duration::from_secs(3)).is_ok()
-                    })
-                    .unwrap_or(false);
-
-                if reachable {
-                    if !json {
-                        ui::check_ok(&format!("{name} endpoint reachable ({endpoint})"));
-                    }
-                    checks.push(serde_json::json!({"check": "network_connectivity", "provider": name, "endpoint": endpoint, "env_var": env_var, "status": "ok"}));
-                } else {
-                    if !json {
-                        ui::check_warn(&format!("{name} endpoint unreachable ({endpoint})"));
-                    }
-                    checks.push(serde_json::json!({"check": "network_connectivity", "provider": name, "endpoint": endpoint, "env_var": env_var, "status": "warn"}));
-                }
-            }
-        }
-    }
-
-    // --- Check 10: Channel token format validation ---
-    if !json {
-        println!("\n  Channel Integrations:");
-    }
-    let channel_keys = [
-        ("TELEGRAM_BOT_TOKEN", "Telegram"),
-        ("DISCORD_BOT_TOKEN", "Discord"),
-        ("SLACK_APP_TOKEN", "Slack App"),
-        ("SLACK_BOT_TOKEN", "Slack Bot"),
-    ];
-    for (env_var, name) in &channel_keys {
-        let set = std::env::var(env_var).is_ok();
-        if set {
-            // Format validation
-            let val = std::env::var(env_var).unwrap_or_default();
-            let format_ok = match *env_var {
-                "TELEGRAM_BOT_TOKEN" => val.contains(':'), // Telegram tokens have format "123456:ABC-DEF..."
-                "DISCORD_BOT_TOKEN" => val.len() > 50,     // Discord tokens are typically 59+ chars
-                "SLACK_APP_TOKEN" => val.starts_with("xapp-"),
-                "SLACK_BOT_TOKEN" => val.starts_with("xoxb-"),
-                _ => true,
-            };
-            if format_ok {
-                if !json {
-                    ui::provider_status(name, env_var, true);
-                }
-            } else if !json {
-                ui::check_warn(&format!("{name} ({env_var}) - unexpected token format"));
-            }
-            checks.push(serde_json::json!({"check": "channel", "name": name, "env_var": env_var, "status": if format_ok { "ok" } else { "warn" }}));
-        } else {
-            if !json {
-                ui::provider_status(name, env_var, false);
-            }
-            checks.push(serde_json::json!({"check": "channel", "name": name, "env_var": env_var, "status": "warn"}));
-        }
-    }
-
-    // --- Check 11: .env keys vs config api_key_env consistency ---
-    {
-        let librefang_dir = cli_librefang_home();
-        let config_path = librefang_dir.join("config.toml");
-        if config_path.exists() {
-            let config_str = std::fs::read_to_string(&config_path).unwrap_or_default();
-            // Look for api_key_env references in config
-            for line in config_str.lines() {
-                let trimmed = line.trim();
-                if let Some(rest) = trimmed.strip_prefix("api_key_env") {
-                    if let Some(val_part) = rest.strip_prefix('=') {
-                        let val = val_part.trim().trim_matches('"');
-                        if !val.is_empty() && std::env::var(val).is_err() {
-                            if !json {
-                                ui::check_warn(&format!(
-                                    "Config references {val} but it is not set in env or .env"
-                                ));
-                            }
-                            checks.push(serde_json::json!({"check": "env_consistency", "status": "warn", "missing_var": val}));
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // --- Check 12: Config deserialization into KernelConfig ---
-    {
-        let librefang_dir = cli_librefang_home();
-        let config_path = librefang_dir.join("config.toml");
-        if config_path.exists() {
-            if !json {
-                println!("\n  Config Validation:");
-            }
-            let config_content = std::fs::read_to_string(&config_path).unwrap_or_default();
-            match toml::from_str::<librefang_types::config::KernelConfig>(&config_content) {
-                Ok(cfg) => {
-                    if !json {
-                        ui::check_ok("Config deserializes into KernelConfig");
-                    }
-                    checks.push(serde_json::json!({"check": "config_deser", "status": "ok"}));
-
-                    // Check exec policy
-                    let mode = format!("{:?}", cfg.exec_policy.mode);
-                    let safe_bins_count = cfg.exec_policy.safe_bins.len();
-                    if !json {
-                        ui::check_ok(&format!(
-                            "Exec policy: mode={mode}, safe_bins={safe_bins_count}"
-                        ));
-                    }
-                    checks.push(serde_json::json!({"check": "exec_policy", "status": "ok", "mode": mode, "safe_bins": safe_bins_count}));
-
-                    // Check includes
-                    if !cfg.include.is_empty() {
-                        let mut include_ok = true;
-                        for inc in &cfg.include {
-                            let inc_path = librefang_dir.join(inc);
-                            if inc_path.exists() {
-                                if !json {
-                                    ui::check_ok(&format!("Include file: {inc}"));
-                                }
-                            } else if repair {
-                                if !json {
-                                    ui::check_warn(&format!("Include file missing: {inc}"));
-                                }
-                                include_ok = false;
-                            } else {
-                                if !json {
-                                    ui::check_fail(&format!("Include file not found: {inc}"));
-                                }
-                                include_ok = false;
-                                all_ok = false;
-                            }
-                        }
-                        checks.push(serde_json::json!({"check": "config_includes", "status": if include_ok { "ok" } else { "fail" }, "count": cfg.include.len()}));
-                    }
-
-                    // Check MCP server configs
-                    if !cfg.mcp_servers.is_empty() {
-                        let mcp_count = cfg.mcp_servers.len();
-                        if !json {
-                            ui::check_ok(&format!("MCP servers configured: {mcp_count}"));
-                        }
-                        for server in &cfg.mcp_servers {
-                            // Validate transport config
-                            let Some(ref transport) = server.transport else {
-                                continue;
-                            };
-                            match transport {
-                                librefang_types::config::McpTransportEntry::Stdio {
-                                    command,
-                                    ..
-                                } => {
-                                    if command.is_empty() {
-                                        if !json {
-                                            ui::check_warn(&format!(
-                                                "MCP server '{}' has empty command",
-                                                server.name
-                                            ));
-                                        }
-                                        checks.push(serde_json::json!({"check": "mcp_server_config", "status": "warn", "name": server.name}));
-                                    }
-                                }
-                                librefang_types::config::McpTransportEntry::Sse { url }
-                                | librefang_types::config::McpTransportEntry::Http { url } => {
-                                    if url.is_empty() {
-                                        if !json {
-                                            ui::check_warn(&format!(
-                                                "MCP server '{}' has empty URL",
-                                                server.name
-                                            ));
-                                        }
-                                        checks.push(serde_json::json!({"check": "mcp_server_config", "status": "warn", "name": server.name}));
-                                    }
-                                }
-                                librefang_types::config::McpTransportEntry::HttpCompat {
-                                    base_url,
-                                    headers,
-                                    tools,
-                                } => {
-                                    if base_url.is_empty() {
-                                        if !json {
-                                            ui::check_warn(&format!(
-                                                "MCP server '{}' has empty base_url",
-                                                server.name
-                                            ));
-                                        }
-                                        checks.push(serde_json::json!({"check": "mcp_server_config", "status": "warn", "name": server.name}));
-                                    }
-                                    if tools.is_empty() {
-                                        if !json {
-                                            ui::check_warn(&format!(
-                                                "MCP server '{}' has no http_compat tools configured",
-                                                server.name
-                                            ));
-                                        }
-                                        checks.push(serde_json::json!({"check": "mcp_server_config", "status": "warn", "name": server.name}));
-                                    }
-                                    if headers.iter().any(|h| h.name.trim().is_empty()) {
-                                        if !json {
-                                            ui::check_warn(&format!(
-                                                "MCP server '{}' has an http_compat header with empty name",
-                                                server.name
-                                            ));
-                                        }
-                                        checks.push(serde_json::json!({"check": "mcp_server_config", "status": "warn", "name": server.name}));
-                                    }
-                                    if headers.iter().any(|h| {
-                                        h.value.as_ref().is_none_or(|value| value.trim().is_empty())
-                                            && h.value_env
-                                                .as_ref()
-                                                .is_none_or(|value| value.trim().is_empty())
-                                    }) {
-                                        if !json {
-                                            ui::check_warn(&format!(
-                                                "MCP server '{}' has an http_compat header without value/value_env",
-                                                server.name
-                                            ));
-                                        }
-                                        checks.push(serde_json::json!({"check": "mcp_server_config", "status": "warn", "name": server.name}));
-                                    }
-                                    if tools.iter().any(|tool| tool.name.trim().is_empty()) {
-                                        if !json {
-                                            ui::check_warn(&format!(
-                                                "MCP server '{}' has an http_compat tool with empty name",
-                                                server.name
-                                            ));
-                                        }
-                                        checks.push(serde_json::json!({"check": "mcp_server_config", "status": "warn", "name": server.name}));
-                                    }
-                                    if tools.iter().any(|tool| tool.path.trim().is_empty()) {
-                                        if !json {
-                                            ui::check_warn(&format!(
-                                                "MCP server '{}' has an http_compat tool with empty path",
-                                                server.name
-                                            ));
-                                        }
-                                        checks.push(serde_json::json!({"check": "mcp_server_config", "status": "warn", "name": server.name}));
-                                    }
-                                }
-                            }
-                        }
-                        checks.push(serde_json::json!({"check": "mcp_servers", "status": "ok", "count": mcp_count}));
-                    }
-                }
-                Err(e) => {
-                    if !json {
-                        ui::check_fail(&format!("Config fails KernelConfig deserialization: {e}"));
-                    }
-                    checks.push(serde_json::json!({"check": "config_deser", "status": "fail", "error": e.to_string()}));
-                    all_ok = false;
-                }
-            }
-        }
-    }
-
-    // --- Check 13: Skill registry health ---
-    {
-        if !json {
-            println!("\n  Skills:");
-        }
-        let skills_dir = cli_librefang_home().join("skills");
-        let mut skill_reg = librefang_skills::registry::SkillRegistry::new(skills_dir.clone());
-        match skill_reg.load_all() {
-            Ok(count) => {
-                if !json {
-                    ui::check_ok(&format!("Skills loaded: {count}"));
-                }
-                checks.push(serde_json::json!({"check": "skills", "status": "ok", "count": count}));
-            }
-            Err(e) => {
-                if !json {
-                    ui::check_warn(&format!("Failed to load skills: {e}"));
-                }
-                checks.push(serde_json::json!({"check": "skills", "status": "warn", "error": e.to_string()}));
-            }
-        }
-
-        // Check for prompt injection issues in skill definitions.
-        // Only flag Critical-severity warnings.
-        let skills = skill_reg.list();
-        let mut injection_warnings = 0;
-        for skill in &skills {
-            if let Some(ref prompt) = skill.manifest.prompt_context {
-                let warnings = librefang_skills::verify::SkillVerifier::scan_prompt_content(prompt);
-                let has_critical = warnings.iter().any(|w| {
-                    matches!(
-                        w.severity,
-                        librefang_skills::verify::WarningSeverity::Critical
-                    )
-                });
-                if has_critical {
-                    injection_warnings += 1;
-                    if !json {
-                        ui::check_warn(&format!(
-                            "Prompt injection warning in skill: {}",
-                            skill.manifest.skill.name
-                        ));
-                    }
-                }
-            }
-        }
-        if injection_warnings > 0 {
-            checks.push(serde_json::json!({"check": "skill_injection_scan", "status": "warn", "warnings": injection_warnings}));
-        } else {
-            if !json {
-                ui::check_ok("All skills pass prompt injection scan");
-            }
-            checks.push(serde_json::json!({"check": "skill_injection_scan", "status": "ok"}));
-        }
-    }
-
-    // --- Check 14: MCP catalog + configured servers ---
-    {
-        if !json {
-            println!("\n  MCP servers:");
-        }
-        let librefang_dir = cli_librefang_home();
-        let mut catalog = librefang_extensions::catalog::McpCatalog::new(&librefang_dir);
-        catalog.load(&librefang_runtime::registry_sync::resolve_home_dir_for_tests());
-        let template_count = catalog.len();
-
-        // Count configured [[mcp_servers]] entries in config.toml (if any).
-        let configured_count = {
-            let config_path = librefang_dir.join("config.toml");
-            if config_path.is_file() {
-                let raw = std::fs::read_to_string(&config_path).unwrap_or_default();
-                toml::from_str::<toml::Value>(&raw)
-                    .ok()
-                    .and_then(|v| v.as_table().cloned())
-                    .and_then(|t| t.get("mcp_servers").cloned())
-                    .and_then(|v| v.as_array().cloned())
-                    .map(|a| a.len())
-                    .unwrap_or(0)
-            } else {
-                0
-            }
-        };
-        if !json {
-            ui::check_ok(&format!("MCP catalog templates: {template_count}"));
-            ui::check_ok(&format!("Configured MCP servers: {configured_count}"));
-        }
-        checks.push(
-            serde_json::json!({"check": "mcp_catalog", "status": "ok", "count": template_count}),
-        );
-        checks.push(serde_json::json!({"check": "mcp_servers_configured", "status": "ok", "count": configured_count}));
-    }
-
-    // --- Check 15: Daemon health detail (if running) ---
-    if let Some(ref base) = find_daemon() {
-        if !json {
-            println!("\n  Daemon Health:");
-        }
-        let client = daemon_client();
-        match client.get(format!("{base}/api/health/detail")).send() {
-            Ok(resp) if resp.status().is_success() => {
-                if let Ok(body) = resp.json::<serde_json::Value>() {
-                    if let Some(agents) = body.get("agent_count").and_then(|v| v.as_u64()) {
-                        if !json {
-                            ui::check_ok(&format!("Running agents: {agents}"));
-                        }
-                        checks.push(serde_json::json!({"check": "daemon_agents", "status": "ok", "count": agents}));
-                    }
-                    if let Some(uptime) = body.get("uptime_secs").and_then(|v| v.as_u64()) {
-                        let hours = uptime / 3600;
-                        let mins = (uptime % 3600) / 60;
-                        if !json {
-                            ui::check_ok(&format!("Daemon uptime: {hours}h {mins}m"));
-                        }
-                        checks.push(serde_json::json!({"check": "daemon_uptime", "status": "ok", "secs": uptime}));
-                    }
-                    if let Some(db_status) = body.get("database").and_then(|v| v.as_str()) {
-                        if db_status == "connected" || db_status == "ok" {
-                            if !json {
-                                ui::check_ok("Database connectivity: OK");
-                            }
-                        } else {
-                            if !json {
-                                ui::check_fail(&format!("Database status: {db_status}"));
-                            }
-                            all_ok = false;
-                        }
-                        checks.push(serde_json::json!({"check": "daemon_db", "status": db_status}));
-                    }
-                }
-            }
-            Ok(resp) => {
-                if !json {
-                    ui::check_warn(&format!("Health detail returned {}", resp.status()));
-                }
-                checks.push(serde_json::json!({"check": "daemon_health", "status": "warn"}));
-            }
-            Err(e) => {
-                if !json {
-                    ui::check_warn(&format!("Failed to query daemon health: {e}"));
-                }
-                checks.push(serde_json::json!({"check": "daemon_health", "status": "warn", "error": e.to_string()}));
-            }
-        }
-
-        // Check skills endpoint
-        match client.get(format!("{base}/api/skills")).send() {
-            Ok(resp) if resp.status().is_success() => {
-                if let Ok(body) = resp.json::<serde_json::Value>() {
-                    if let Some(arr) = body
-                        .get("skills")
-                        .and_then(|v| v.as_array())
-                        .or_else(|| body.as_array())
-                    {
-                        if !json {
-                            ui::check_ok(&format!("Skills loaded in daemon: {}", arr.len()));
-                        }
-                        checks.push(serde_json::json!({"check": "daemon_skills", "status": "ok", "count": arr.len()}));
-                    }
-                }
-            }
-            _ => {}
-        }
-
-        // Check MCP servers endpoint
-        match client.get(format!("{base}/api/mcp/servers")).send() {
-            Ok(resp) if resp.status().is_success() => {
-                if let Ok(body) = resp.json::<serde_json::Value>() {
-                    if let Some(arr) = body
-                        .get("configured")
-                        .and_then(|v| v.as_array())
-                        .or_else(|| body.as_array())
-                    {
-                        let connected = arr
-                            .iter()
-                            .filter(|s| {
-                                s.get("connected")
-                                    .and_then(|v| v.as_bool())
-                                    .unwrap_or(false)
-                            })
-                            .count();
-                        if !json {
-                            ui::check_ok(&format!(
-                                "MCP servers: {} configured, {} connected",
-                                arr.len(),
-                                connected
-                            ));
-                        }
-                        checks.push(serde_json::json!({"check": "daemon_mcp", "status": "ok", "configured": arr.len(), "connected": connected}));
-                    }
-                }
-            }
-            _ => {}
-        }
-
-        // Check MCP health endpoint
-        match client.get(format!("{base}/api/mcp/health")).send() {
-            Ok(resp) if resp.status().is_success() => {
-                if let Ok(body) = resp.json::<serde_json::Value>() {
-                    let entries = body.get("health").and_then(|h| h.as_array());
-                    if let Some(arr) = entries {
-                        let healthy = arr
-                            .iter()
-                            .filter(|v| {
-                                v.get("status")
-                                    .and_then(|s| s.as_str())
-                                    .map(|s| s.eq_ignore_ascii_case("ready"))
-                                    .unwrap_or(false)
-                            })
-                            .count();
-                        let total = arr.len();
-                        if healthy == total {
-                            if !json {
-                                ui::check_ok(&format!(
-                                    "MCP server health: {healthy}/{total} healthy"
-                                ));
-                            }
-                        } else if !json {
-                            ui::check_warn(&format!(
-                                "MCP server health: {healthy}/{total} healthy"
-                            ));
-                        }
-                        checks.push(serde_json::json!({"check": "mcp_health", "status": if healthy == total { "ok" } else { "warn" }, "healthy": healthy, "total": total}));
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    if !json {
-        println!();
-    }
-    match std::process::Command::new("rustc")
-        .arg("--version")
-        .output()
-    {
-        Ok(output) => {
-            let version = String::from_utf8_lossy(&output.stdout).trim().to_string();
-            if !json {
-                ui::check_ok(&format!("Rust: {version}"));
-            }
-            checks.push(serde_json::json!({"check": "rust", "status": "ok", "version": version}));
-        }
-        Err(_) => {
-            if !json {
-                ui::check_fail("Rust toolchain not found");
-            }
-            checks.push(serde_json::json!({"check": "rust", "status": "fail"}));
-            all_ok = false;
-        }
-    }
-
-    // Python runtime check
-    match std::process::Command::new("python3")
-        .arg("--version")
-        .output()
-    {
-        Ok(output) if output.status.success() => {
-            let version = String::from_utf8_lossy(&output.stdout).trim().to_string();
-            if !json {
-                ui::check_ok(&format!("Python: {version}"));
-            }
-            checks.push(serde_json::json!({"check": "python", "status": "ok", "version": version}));
-        }
-        _ => {
-            // Try `python` instead
-            match std::process::Command::new("python")
-                .arg("--version")
-                .output()
-            {
-                Ok(output) if output.status.success() => {
-                    let version = String::from_utf8_lossy(&output.stdout).trim().to_string();
-                    if !json {
-                        ui::check_ok(&format!("Python: {version}"));
-                    }
-                    checks.push(
-                        serde_json::json!({"check": "python", "status": "ok", "version": version}),
-                    );
-                }
-                _ => {
-                    if !json {
-                        ui::check_warn("Python not found (needed for Python skill runtime)");
-                    }
-                    checks.push(serde_json::json!({"check": "python", "status": "warn"}));
-                }
-            }
-        }
-    }
-
-    // Node.js runtime check
-    match std::process::Command::new("node").arg("--version").output() {
-        Ok(output) if output.status.success() => {
-            let version = String::from_utf8_lossy(&output.stdout).trim().to_string();
-            if !json {
-                ui::check_ok(&format!("Node.js: {version}"));
-            }
-            checks.push(serde_json::json!({"check": "node", "status": "ok", "version": version}));
-        }
-        _ => {
-            if !json {
-                ui::check_warn("Node.js not found (needed for Node skill runtime)");
-            }
-            checks.push(serde_json::json!({"check": "node", "status": "warn"}));
-        }
-    }
-
-    // Framework-based audit checks (see crates/librefang-cli/src/doctor.rs).
-    // Each check is its own struct, registered in `doctor::registered_checks`.
-    // Migrating the legacy inline checks above into this framework can happen
-    // incrementally — adding a new check is one struct + one registry entry,
-    // no edits to this function.
-    {
-        let ctx = doctor::AuditContext {
-            librefang_home: cli_librefang_home(),
-        };
-        for result in doctor::run_all(&ctx) {
-            if !json {
-                match result.severity {
-                    doctor::Severity::Pass | doctor::Severity::Info => {
-                        ui::check_ok(&result.summary);
-                    }
-                    doctor::Severity::Warn => {
-                        ui::check_warn(&result.summary);
-                        if let Some(hint) = &result.hint {
-                            ui::hint(hint);
-                        }
-                    }
-                    doctor::Severity::Error => {
-                        ui::check_fail(&result.summary);
-                        if let Some(hint) = &result.hint {
-                            ui::hint(hint);
-                        }
-                    }
-                }
-            }
-            let mut entry = serde_json::json!({
-                "check": result.name,
-                "status": result.severity.as_str(),
-                "summary": result.summary,
-            });
-            if let Some(h) = &result.hint {
-                entry["hint"] = serde_json::Value::String(h.clone());
-            }
-            checks.push(entry);
-            if matches!(result.severity, doctor::Severity::Error) {
-                all_ok = false;
-            }
-        }
-    }
-
-    if json {
-        write_stdout_safe(
-            &serde_json::to_string_pretty(&serde_json::json!({
-                "all_ok": all_ok,
-                "checks": checks,
-            }))
-            .unwrap_or_default(),
-        );
-    } else {
-        println!();
-        if all_ok {
-            ui::success(&i18n::t("doctor-all-passed"));
-            ui::hint(&i18n::t("hint-start-daemon-cmd"));
-        } else if repaired {
-            ui::success(&i18n::t("doctor-repairs-applied"));
-        } else {
-            ui::error(&i18n::t("doctor-some-failed"));
-            if !repair {
-                ui::hint(&i18n::t("hint-doctor-repair"));
-            }
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Dashboard command
-// ---------------------------------------------------------------------------
-
-fn cmd_dashboard() {
-    let base = if let Some(url) = find_daemon() {
-        url
-    } else {
-        // Auto-start the daemon
-        ui::hint(&i18n::t("daemon-no-running-auto"));
-        match start_daemon_background() {
-            Ok(url) => {
-                ui::success(&i18n::t("daemon-started"));
-                url
-            }
-            Err(e) => {
-                ui::error_with_fix(
-                    &i18n::t_args("daemon-start-fail", &[("error", &e.to_string())]),
-                    &i18n::t("daemon-start-fail-fix"),
-                );
-                std::process::exit(1);
-            }
-        }
-    };
-
-    let url = format!("{base}/");
-    ui::success(&i18n::t_args("dashboard-opening", &[("url", &url)]));
-    if copy_to_clipboard(&url) {
-        ui::hint(&i18n::t("hint-url-copied"));
-    }
-    if !open_in_browser(&url) {
-        ui::hint(&i18n::t_args(
-            "hint-could-not-open-browser-visit",
-            &[("url", &url)],
-        ));
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Shell completion command
-// ---------------------------------------------------------------------------
-
-fn cmd_completion(shell: clap_complete::Shell) {
-    use clap::CommandFactory;
-    let mut cmd = Cli::command();
-    clap_complete::generate(shell, &mut cmd, "librefang", &mut std::io::stdout());
-}
-
-// ---------------------------------------------------------------------------
-// Migrate command
-// ---------------------------------------------------------------------------
-
-fn cmd_migrate(args: MigrateArgs) {
-    let source = match args.from {
-        MigrateSourceArg::Openclaw => librefang_import::MigrateSource::OpenClaw,
-        MigrateSourceArg::Langchain => librefang_import::MigrateSource::LangChain,
-        MigrateSourceArg::Autogpt => librefang_import::MigrateSource::AutoGpt,
-        MigrateSourceArg::Openfang => librefang_import::MigrateSource::OpenFang,
-    };
-
-    let source_dir = args.source_dir.unwrap_or_else(|| {
-        let home = dirs::home_dir().unwrap_or_else(|| {
-            eprintln!("Error: Could not determine home directory");
-            std::process::exit(1);
-        });
-        match source {
-            librefang_import::MigrateSource::OpenClaw => home.join(".openclaw"),
-            librefang_import::MigrateSource::LangChain => home.join(".langchain"),
-            librefang_import::MigrateSource::AutoGpt => home.join("Auto-GPT"),
-            librefang_import::MigrateSource::OpenFang => home.join(".openfang"),
-        }
-    });
-
-    let target_dir = cli_librefang_home();
-
-    println!("Migrating from {} ({})...", source, source_dir.display());
-    if args.dry_run {
-        println!("  (dry run — no changes will be made)\n");
-    }
-
-    let options = librefang_import::MigrateOptions {
-        source,
-        source_dir,
-        target_dir,
-        dry_run: args.dry_run,
-    };
-
-    let mut sp = progress::auto("Running migration", None);
-    match librefang_import::run_migration(&options) {
-        Ok(report) => {
-            sp.finish("Migration complete");
-            report.print_summary();
-
-            // Save migration report
-            if !args.dry_run {
-                let report_path = options.target_dir.join("migration_report.md");
-                if let Err(e) = std::fs::write(&report_path, report.to_markdown()) {
-                    eprintln!("Warning: Could not save migration report: {e}");
-                } else {
-                    println!("\n  Report saved to: {}", report_path.display());
-                }
-            }
-        }
-        Err(e) => {
-            sp.finish_with_failure(&format!("Migration failed: {e}"));
-            std::process::exit(1);
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Scaffold commands (librefang new skill/integration)
-// ---------------------------------------------------------------------------
-
-fn cmd_scaffold(kind: ScaffoldKind) {
-    let cwd = std::env::current_dir().unwrap_or_default();
-    let result = match kind {
-        ScaffoldKind::Skill => {
-            librefang_extensions::installer::scaffold_skill(&cwd.join("my-skill"))
-        }
-        ScaffoldKind::Mcp => {
-            librefang_extensions::installer::scaffold_integration(&cwd.join("my-mcp"))
-        }
-    };
-    match result {
-        Ok(msg) => ui::success(&msg),
-        Err(e) => {
-            ui::error(&e.to_string());
-            std::process::exit(1);
-        }
-    }
-}
-
-fn cmd_security_status(json: bool) {
-    let base = require_daemon("security status");
-    let client = daemon_client();
-    let body = daemon_json(client.get(format!("{base}/api/health/detail")).send());
-    if json {
-        let data = serde_json::json!({
-            "audit_trail": "merkle_hash_chain_sha256",
-            "taint_tracking": "information_flow_labels",
-            "wasm_sandbox": "dual_metering_fuel_epoch",
-            "wire_protocol": "ofp_hmac_sha256_mutual_auth",
-            "api_keys": "zeroizing_auto_wipe",
-            "manifests": "ed25519_signed",
-            "agent_count": body.get("agent_count").and_then(|v| v.as_u64()),
-        });
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&data).unwrap_or_default()
-        );
-        return;
-    }
-    ui::section(&i18n::t("section-security-status"));
-    ui::blank();
-    ui::kv(&i18n::t("label-audit-trail"), &i18n::t("value-audit-trail"));
-    ui::kv(
-        &i18n::t("label-taint-tracking"),
-        &i18n::t("value-taint-tracking"),
-    );
-    ui::kv(
-        &i18n::t("label-wasm-sandbox"),
-        &i18n::t("value-wasm-sandbox"),
-    );
-    ui::kv(
-        &i18n::t("label-wire-protocol"),
-        &i18n::t("value-wire-protocol"),
-    );
-    ui::kv(&i18n::t("label-api-keys"), &i18n::t("value-api-keys"));
-    ui::kv(&i18n::t("label-manifests"), &i18n::t("value-manifests"));
-    if let Some(agents) = body.get("agent_count").and_then(|v| v.as_u64()) {
-        ui::kv(&i18n::t("label-active-agents"), &agents.to_string());
-    }
-}
-
-fn cmd_security_audit(limit: usize, json: bool) {
-    let base = require_daemon("security audit");
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .get(format!("{base}/api/audit/recent?limit={limit}"))
-            .send(),
-    );
-    if json {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-        return;
-    }
-    if let Some(arr) = body
-        .get("entries")
-        .and_then(|v| v.as_array())
-        .or_else(|| body.as_array())
-    {
-        if arr.is_empty() {
-            println!("No audit entries.");
-            return;
-        }
-        let mut t = crate::table::Table::new(&["TIMESTAMP", "AGENT", "TYPE", "EVENT"]);
-        for entry in arr {
-            let agent_id = entry["agent_id"].as_str().unwrap_or("");
-            let agent_col = if agent_id.len() > 16 {
-                &agent_id[..16]
-            } else if agent_id.is_empty() {
-                entry["agent_name"].as_str().unwrap_or("?")
-            } else {
-                agent_id
-            };
-            t.add_row(&[
-                entry["timestamp"].as_str().unwrap_or("?"),
-                agent_col,
-                entry["action"]
-                    .as_str()
-                    .or_else(|| entry["event_type"].as_str())
-                    .unwrap_or("?"),
-                entry["detail"]
-                    .as_str()
-                    .or_else(|| entry["description"].as_str())
-                    .unwrap_or(""),
-            ]);
-        }
-        t.print();
-    } else {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-    }
-}
-
-fn cmd_security_verify() {
-    let base = require_daemon("security verify");
-    let client = daemon_client();
-    let body = daemon_json(client.get(format!("{base}/api/audit/verify")).send());
-    if body["valid"].as_bool().unwrap_or(false) {
-        ui::success(&i18n::t("audit-verified"));
-    } else {
-        ui::error(&i18n::t("audit-failed"));
-        if let Some(msg) = body["error"].as_str() {
-            ui::hint(msg);
-        }
-        std::process::exit(1);
-    }
-}
-
-/// Destructively reset the local audit trail.
-///
-/// Truncates `audit_entries` in SQLite and removes the anchor file so the
-/// next daemon boot seeds a fresh Merkle chain. Refuses to run while the
-/// daemon holds the DB (SQLite WAL mode + writer lock) and without
-/// `--confirm`.
-fn cmd_audit_reset(config: Option<PathBuf>, confirm: bool) {
-    let daemon = daemon_config_context(config.as_deref());
-    // `load_config` already eprintln!s the underlying parse / deserialize
-    // error (see #5186); printing it again here would double the message.
-    let kernel_config = match load_config(config.as_deref()) {
-        Ok(cfg) => cfg,
-        Err(_) => std::process::exit(1),
-    };
-
-    let db_path = kernel_config
-        .memory
-        .sqlite_path
-        .clone()
-        .unwrap_or_else(|| kernel_config.data_dir.join("librefang.db"));
-
-    let anchor_path = match kernel_config.audit.anchor_path.as_ref() {
-        Some(p) if p.is_absolute() => p.clone(),
-        Some(p) => kernel_config.data_dir.join(p),
-        None => kernel_config.data_dir.join("audit.anchor"),
-    };
-
-    if !confirm {
-        ui::error("audit reset is destructive — re-run with `--confirm` to proceed");
-        ui::blank();
-        println!("  Would:");
-        println!(
-            "    1. DELETE all rows from `audit_entries` in {}",
-            db_path.display()
-        );
-        println!("    2. Remove anchor file {}", anchor_path.display());
-        println!("  The Merkle chain will restart from the next audit event.");
-        std::process::exit(1);
-    }
-
-    // Refuse if daemon is running — SQLite writer lock would block or corrupt.
-    if let Some(base) = find_daemon_in_home(&daemon.home_dir) {
-        ui::error_with_fix(
-            &format!("daemon is running at {base}; refusing to touch the audit database"),
-            "stop the daemon first: `librefang stop`",
-        );
-        std::process::exit(1);
-    }
-
-    if !db_path.exists() {
-        ui::error(&format!("database not found at {}", db_path.display()));
-        std::process::exit(1);
-    }
-
-    let conn = match rusqlite::Connection::open(&db_path) {
-        Ok(c) => c,
-        Err(e) => {
-            ui::error(&format!("failed to open {}: {e}", db_path.display()));
-            std::process::exit(1);
-        }
-    };
-
-    let rows_before: i64 = conn
-        .query_row("SELECT COUNT(*) FROM audit_entries", [], |r| r.get(0))
-        .unwrap_or(0);
-
-    // Remove the anchor FIRST. If the subsequent DB truncation then fails,
-    // the next daemon boot sees `read_anchor = None` and re-seeds from the
-    // current DB tip — a consistent (if still broken) state the user can
-    // retry. The reverse order (DB first, anchor second) would instead
-    // leave an empty table alongside a stale anchor, which produces a
-    // fresh MISMATCH error the user didn't have before calling reset.
-    let anchor_removed = if anchor_path.exists() {
-        match std::fs::remove_file(&anchor_path) {
-            Ok(()) => true,
-            Err(e) => {
-                ui::error(&format!(
-                    "failed to remove anchor {}: {e}",
-                    anchor_path.display()
-                ));
-                std::process::exit(1);
-            }
-        }
-    } else {
-        false
-    };
-
-    if let Err(e) = conn.execute("DELETE FROM audit_entries", []) {
-        ui::error(&format!("failed to truncate audit_entries: {e}"));
-        std::process::exit(1);
-    }
-    drop(conn);
-    // `seq` is `INTEGER PRIMARY KEY` without AUTOINCREMENT, so the next
-    // insert after an empty table naturally gets seq = 1. No sqlite_sequence
-    // fiddling needed.
-
-    ui::success(&format!(
-        "Audit trail reset: removed {rows_before} row(s) from audit_entries{}.",
-        if anchor_removed {
-            format!(", deleted anchor at {}", anchor_path.display())
-        } else {
-            " (no anchor file to remove)".to_string()
-        }
-    ));
-    ui::hint("The next daemon boot will seed a fresh Merkle chain from the current tip.");
-}
-
-fn cmd_memory_list(agent: &str, json: bool) {
-    let base = require_daemon("memory list");
-    let agent = resolve_agent_id(&base, agent);
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .get(format!("{base}/api/memory/agents/{agent}/kv"))
-            .send(),
-    );
-    if json {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-        return;
-    }
-    if let Some(arr) = body
-        .get("kv_pairs")
-        .and_then(|v| v.as_array())
-        .or_else(|| body.as_array())
-    {
-        if arr.is_empty() {
-            println!("No memory entries for agent '{agent}'.");
-            return;
-        }
-        let mut t = crate::table::Table::new(&["KEY", "VALUE"]);
-        for kv in arr {
-            t.add_row(&[
-                kv["key"].as_str().unwrap_or("?"),
-                &kv["value"]
-                    .as_str()
-                    .unwrap_or("")
-                    .chars()
-                    .take(50)
-                    .collect::<String>(),
-            ]);
-        }
-        t.print();
-    } else {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-    }
-}
-
-fn cmd_memory_get(agent: &str, key: &str, json: bool) {
-    let base = require_daemon("memory get");
-    let agent = resolve_agent_id(&base, agent);
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .get(format!("{base}/api/memory/agents/{agent}/kv/{key}"))
-            .send(),
-    );
-    if json {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-        return;
-    }
-    if let Some(val) = body["value"].as_str() {
-        println!("{val}");
-    } else {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-    }
-}
-
-fn cmd_memory_set(agent: &str, key: &str, value: &str) {
-    let base = require_daemon("memory set");
-    let agent = resolve_agent_id(&base, agent);
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .put(format!("{base}/api/memory/agents/{agent}/kv/{key}"))
-            .json(&serde_json::json!({"value": value}))
-            .send(),
-    );
-    if body.get("error").is_some() {
-        ui::error(&i18n::t_args(
-            "memory-set-failed",
-            &[("error", body["error"].as_str().unwrap_or("?"))],
-        ));
-    } else {
-        ui::success(&i18n::t_args(
-            "memory-set",
-            &[("key", key), ("agent", &agent)],
-        ));
-    }
-}
-
-fn cmd_memory_delete(agent: &str, key: &str) {
-    let base = require_daemon("memory delete");
-    let agent = resolve_agent_id(&base, agent);
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .delete(format!("{base}/api/memory/agents/{agent}/kv/{key}"))
-            .send(),
-    );
-    if body.get("error").is_some() {
-        ui::error(&i18n::t_args(
-            "memory-delete-failed",
-            &[("error", body["error"].as_str().unwrap_or("?"))],
-        ));
-    } else {
-        ui::success(&i18n::t_args(
-            "memory-deleted",
-            &[("key", key), ("agent", &agent)],
-        ));
-    }
-}
-
-fn cmd_devices_list(json: bool) {
-    let base = require_daemon("devices list");
-    let client = daemon_client();
-    let body = daemon_json(client.get(format!("{base}/api/pairing/devices")).send());
-    if json {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-        return;
-    }
-    if let Some(arr) = body.as_array() {
-        if arr.is_empty() {
-            println!("No paired devices.");
-            return;
-        }
-        let mut t = crate::table::Table::new(&["ID", "NAME", "LAST SEEN"]);
-        for d in arr {
-            t.add_row(&[
-                d["id"].as_str().unwrap_or("?"),
-                d["name"].as_str().unwrap_or("?"),
-                d["last_seen"].as_str().unwrap_or("?"),
-            ]);
-        }
-        t.print();
-    } else {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-    }
-}
-
-fn cmd_devices_pair() {
-    let base = require_daemon("qr");
-    let client = daemon_client();
-    let body = daemon_json(client.post(format!("{base}/api/pairing/request")).send());
-    if let Some(qr) = body["qr_data"].as_str() {
-        ui::section(&i18n::t("section-device-pairing"));
-        ui::blank();
-        // Render a simple text-based QR representation
-        println!("  {}", i18n::t("device-scan-qr"));
-        ui::blank();
-        println!("  {qr}");
-        ui::blank();
-        if let Some(code) = body["pairing_code"].as_str() {
-            ui::kv(&i18n::t("label-pairing-code"), code);
-        }
-        if let Some(expires) = body["expires_at"].as_str() {
-            ui::kv(&i18n::t("label-expires"), expires);
-        }
-    } else {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-    }
-}
-
-fn cmd_devices_remove(id: &str) {
-    let base = require_daemon("devices remove");
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .delete(format!("{base}/api/pairing/devices/{id}"))
-            .send(),
-    );
-    if body.get("error").is_some() {
-        ui::error(&i18n::t_args(
-            "device-remove-failed",
-            &[("error", body["error"].as_str().unwrap_or("?"))],
-        ));
-    } else {
-        ui::success(&i18n::t_args("device-removed", &[("id", id)]));
-    }
-}
-
-fn cmd_webhooks_list(json: bool) {
-    let base = require_daemon("webhooks list");
-    let client = daemon_client();
-    let body = daemon_json(client.get(format!("{base}/api/webhooks")).send());
-    if json {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-        return;
-    }
-    if let Some(arr) = body
-        .get("webhooks")
-        .and_then(|v| v.as_array())
-        .or_else(|| body.as_array())
-    {
-        if arr.is_empty() {
-            println!("No webhooks configured.");
-            return;
-        }
-        let mut t = crate::table::Table::new(&["ID", "NAME", "ENABLED", "URL"]);
-        for w in arr {
-            t.add_row(&[
-                w["id"].as_str().unwrap_or("?"),
-                w["name"].as_str().unwrap_or("?"),
-                if w["enabled"].as_bool().unwrap_or(false) {
-                    "yes"
-                } else {
-                    "no"
-                },
-                w["url"].as_str().unwrap_or(""),
-            ]);
-        }
-        t.print();
-    } else {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-    }
-}
-
-fn cmd_webhooks_create(agent: &str, url: &str) {
-    let base = require_daemon("webhooks create");
-    let agent = resolve_agent_id(&base, agent);
-    let client = daemon_client();
-
-    // Derive a name from the URL hostname
-    let name = reqwest::Url::parse(url)
-        .ok()
-        .and_then(|u| u.host_str().map(|h| h.to_string()))
-        .unwrap_or_else(|| "webhook".to_string());
-
-    let body = daemon_json(
-        client
-            .post(format!("{base}/api/webhooks"))
-            .json(&serde_json::json!({
-                "name": format!("{agent}-{name}"),
-                "url": url,
-                "events": ["all"],
-            }))
-            .send(),
-    );
-    if let Some(id) = body["id"].as_str() {
-        ui::success(&i18n::t_args("webhook-created", &[("id", id)]));
-    } else {
-        ui::error(&i18n::t_args(
-            "webhook-create-failed",
-            &[("error", body["error"].as_str().unwrap_or("?"))],
-        ));
-    }
-}
-
-fn cmd_webhooks_delete(id: &str) {
-    let base = require_daemon("webhooks delete");
-    let client = daemon_client();
-    let body = daemon_json(client.delete(format!("{base}/api/webhooks/{id}")).send());
-    if body.get("error").is_some() {
-        ui::error(&i18n::t_args(
-            "webhook-delete-failed",
-            &[("error", body["error"].as_str().unwrap_or("?"))],
-        ));
-    } else {
-        ui::success(&i18n::t_args("webhook-deleted", &[("id", id)]));
-    }
-}
-
-fn cmd_webhooks_test(id: &str) {
-    let base = require_daemon("webhooks test");
-    let client = daemon_client();
-    let body = daemon_json(client.post(format!("{base}/api/webhooks/{id}/test")).send());
-    if body["success"].as_bool().unwrap_or(false) {
-        ui::success(&i18n::t_args("webhook-test-ok", &[("id", id)]));
-    } else {
-        ui::error(&i18n::t_args(
-            "webhook-test-failed",
-            &[("error", body["error"].as_str().unwrap_or("?"))],
-        ));
-    }
-}
-
-fn cmd_system_info(json: bool) {
-    if let Some(base) = find_daemon() {
-        let client = daemon_client();
-        let body = daemon_json(client.get(format!("{base}/api/status")).send());
-        if json {
-            let mut data = body.clone();
-            if let Some(obj) = data.as_object_mut() {
-                obj.insert(
-                    "version".to_string(),
-                    serde_json::json!(env!("CARGO_PKG_VERSION")),
-                );
-                obj.insert("api_url".to_string(), serde_json::json!(base));
-            }
-            println!(
-                "{}",
-                serde_json::to_string_pretty(&data).unwrap_or_default()
-            );
-            return;
-        }
-        ui::section(&i18n::t("section-system-info"));
-        ui::blank();
-        ui::kv(&i18n::t("label-version"), env!("CARGO_PKG_VERSION"));
-        ui::kv(
-            &i18n::t("label-status"),
-            body["status"].as_str().unwrap_or("?"),
-        );
-        ui::kv(
-            &i18n::t("label-agents"),
-            &body["agent_count"].as_u64().unwrap_or(0).to_string(),
-        );
-        ui::kv(
-            &i18n::t("label-provider"),
-            body["default_provider"].as_str().unwrap_or("?"),
-        );
-        ui::kv(
-            &i18n::t("label-model"),
-            body["default_model"].as_str().unwrap_or("?"),
-        );
-        ui::kv(&i18n::t("label-api"), &base);
-        ui::kv(
-            &i18n::t("label-data-dir"),
-            body["data_dir"].as_str().unwrap_or("?"),
-        );
-        ui::kv(
-            &i18n::t("label-uptime"),
-            &format!("{}s", body["uptime_seconds"].as_u64().unwrap_or(0)),
-        );
-    } else {
-        if json {
-            println!(
-                "{}",
-                serde_json::json!({
-                    "version": env!("CARGO_PKG_VERSION"),
-                    "daemon": "not_running",
-                })
-            );
-            return;
-        }
-        ui::section(&i18n::t("section-system-info"));
-        ui::blank();
-        ui::kv(&i18n::t("label-version"), env!("CARGO_PKG_VERSION"));
-        ui::kv_warn(
-            &i18n::t("label-daemon"),
-            &i18n::t("label-daemon-not-running"),
-        );
-        ui::hint(&i18n::t("hint-start-daemon"));
-    }
-}
-
-fn cmd_system_version(json: bool) {
-    if json {
-        println!(
-            "{}",
-            serde_json::json!({"version": env!("CARGO_PKG_VERSION")})
-        );
-        return;
-    }
-    println!("librefang {}", env!("CARGO_PKG_VERSION"));
-}
-
-// ---------------------------------------------------------------------------
-// Service management (boot auto-start)
-// ---------------------------------------------------------------------------
-
-/// Resolve the absolute path to the current librefang binary.
-fn resolve_binary_path() -> std::path::PathBuf {
-    std::env::current_exe()
-        .unwrap_or_else(|_| std::path::PathBuf::from("librefang"))
-        .canonicalize()
-        .unwrap_or_else(|_| std::env::current_exe().unwrap_or_else(|_| "librefang".into()))
-}
-
-fn cmd_service_install() {
-    // Warn if running as root — the service would be installed for root, not
-    // the actual user. This catches `sudo librefang service install` mistakes.
-    #[cfg(unix)]
-    {
-        // SAFETY: geteuid() is always safe to call.
-        if unsafe { libc::geteuid() } == 0 {
-            ui::error(
-                "Running as root — the service will be installed for the root account, \
-                 not your user. Run without sudo instead.",
-            );
-            std::process::exit(1);
-        }
-    }
-
-    let binary = resolve_binary_path();
-
-    #[cfg(any(target_os = "linux", target_os = "macos"))]
-    let librefang_home = cli_librefang_home();
-
-    #[cfg(target_os = "linux")]
-    {
-        service_install_linux(&binary, &librefang_home);
-    }
-    #[cfg(target_os = "macos")]
-    {
-        service_install_macos(&binary, &librefang_home);
-    }
-    #[cfg(windows)]
-    {
-        service_install_windows(&binary);
-    }
-    #[cfg(not(any(target_os = "linux", target_os = "macos", windows)))]
-    {
-        let _ = &binary;
-        ui::error("Auto-start service is not supported on this platform.");
-    }
-}
-
-#[cfg(target_os = "linux")]
-fn service_install_linux(binary: &std::path::Path, librefang_home: &std::path::Path) {
-    let home = match dirs::home_dir() {
-        Some(h) => h,
-        None => {
-            ui::error("Cannot determine home directory.");
-            return;
-        }
-    };
-    let service_dir = home.join(".config/systemd/user");
-    if let Err(e) = std::fs::create_dir_all(&service_dir) {
-        ui::error(&format!("Failed to create {}: {e}", service_dir.display()));
-        return;
-    }
-
-    let unit = format!(
-        "[Unit]\n\
-         Description=LibreFang Agent OS Daemon\n\
-         Documentation=https://librefang.ai\n\
-         After=network-online.target\n\
-         Wants=network-online.target\n\
-         \n\
-         [Service]\n\
-         Type=simple\n\
-         ExecStart={binary} start --foreground\n\
-         Restart=on-failure\n\
-         RestartSec=5\n\
-         WorkingDirectory={home}\n\
-         EnvironmentFile=-{home}/env\n\
-         EnvironmentFile=-{home}/secrets.env\n\
-         \n\
-         [Install]\n\
-         WantedBy=default.target\n",
-        binary = binary.display(),
-        home = librefang_home.display(),
-    );
-
-    let service_path = service_dir.join("librefang.service");
-    if let Err(e) = std::fs::write(&service_path, &unit) {
-        ui::error(&format!("Failed to write {}: {e}", service_path.display()));
-        return;
-    }
-    ui::success(&format!("Wrote {}", service_path.display()));
-
-    // Reload and enable
-    let reload = std::process::Command::new("systemctl")
-        .args(["--user", "daemon-reload"])
-        .output();
-    if let Ok(o) = &reload {
-        if !o.status.success() {
-            ui::error("systemctl --user daemon-reload failed");
-            return;
-        }
-    }
-    let enable = std::process::Command::new("systemctl")
-        .args(["--user", "enable", "librefang.service"])
-        .output();
-    match enable {
-        Ok(o) if o.status.success() => {
-            ui::success("Service enabled (will start on next login)");
-            ui::hint("Start now with: systemctl --user start librefang.service");
-            // Enable lingering so the user service runs without an active login session
-            ui::hint("For headless servers, also run: loginctl enable-linger");
-        }
-        _ => ui::error("systemctl --user enable librefang.service failed"),
-    }
-}
-
-#[cfg(target_os = "macos")]
-fn service_install_macos(binary: &std::path::Path, librefang_home: &std::path::Path) {
-    let home = match dirs::home_dir() {
-        Some(h) => h,
-        None => {
-            ui::error("Cannot determine home directory.");
-            return;
-        }
-    };
-    let agents_dir = home.join("Library/LaunchAgents");
-    if let Err(e) = std::fs::create_dir_all(&agents_dir) {
-        ui::error(&format!("Failed to create {}: {e}", agents_dir.display()));
-        return;
-    }
-
-    let plist = format!(
-        r#"<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN"
-  "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-    <key>Label</key>
-    <string>ai.librefang.daemon</string>
-    <key>ProgramArguments</key>
-    <array>
-        <string>{binary}</string>
-        <string>start</string>
-        <string>--foreground</string>
-    </array>
-    <key>RunAtLoad</key>
-    <true/>
-    <key>KeepAlive</key>
-    <true/>
-    <key>WorkingDirectory</key>
-    <string>{home}</string>
-    <key>StandardOutPath</key>
-    <string>{home}/daemon.log</string>
-    <key>StandardErrorPath</key>
-    <string>{home}/daemon.log</string>
-</dict>
-</plist>
-"#,
-        binary = binary.display(),
-        home = librefang_home.display(),
-    );
-
-    let plist_path = agents_dir.join("ai.librefang.daemon.plist");
-
-    // Unload existing service first (if any) to avoid launchctl errors
-    if plist_path.exists() {
-        let _ = std::process::Command::new("launchctl")
-            .args(["unload", &plist_path.to_string_lossy()])
-            .output();
-    }
-
-    if let Err(e) = std::fs::write(&plist_path, &plist) {
-        ui::error(&format!("Failed to write {}: {e}", plist_path.display()));
-        return;
-    }
-    ui::success(&format!("Wrote {}", plist_path.display()));
-
-    let load = std::process::Command::new("launchctl")
-        .args(["load", &plist_path.to_string_lossy()])
-        .output();
-    match load {
-        Ok(o) if o.status.success() => {
-            ui::success("LaunchAgent loaded (will start on login and now)");
-        }
-        Ok(o) => {
-            let stderr = String::from_utf8_lossy(&o.stderr);
-            ui::error(&format!("launchctl load failed: {stderr}"));
-        }
-        Err(e) => ui::error(&format!("Failed to run launchctl: {e}")),
-    }
-}
-
-#[cfg(windows)]
-fn service_install_windows(binary: &std::path::Path) {
-    let value = format!("\"{}\" start", binary.display());
-    let output = std::process::Command::new("reg")
-        .args([
-            "add",
-            r"HKCU\Software\Microsoft\Windows\CurrentVersion\Run",
-            "/v",
-            "LibreFang",
-            "/t",
-            "REG_SZ",
-            "/d",
-            &value,
-            "/f",
-        ])
-        .output();
-    match output {
-        Ok(o) if o.status.success() => {
-            ui::success("Added to Windows startup (HKCU\\...\\Run)");
-        }
-        Ok(o) => {
-            let stderr = String::from_utf8_lossy(&o.stderr);
-            ui::error(&format!("Failed to write registry: {stderr}"));
-        }
-        Err(e) => ui::error(&format!("Failed to run reg.exe: {e}")),
-    }
-}
-
-fn cmd_service_uninstall() {
-    #[cfg(target_os = "linux")]
-    {
-        let home = dirs::home_dir().unwrap_or_default();
-        let service_path = home.join(".config/systemd/user/librefang.service");
-        if service_path.exists() {
-            let _ = std::process::Command::new("systemctl")
-                .args(["--user", "disable", "--now", "librefang.service"])
-                .output();
-            match std::fs::remove_file(&service_path) {
-                Ok(()) => {
-                    let _ = std::process::Command::new("systemctl")
-                        .args(["--user", "daemon-reload"])
-                        .output();
-                    ui::success("Removed systemd user service");
-                }
-                Err(e) => ui::error(&format!("Failed to remove service file: {e}")),
-            }
-        } else {
-            ui::hint("No systemd user service found — nothing to remove.");
-        }
-    }
-    #[cfg(target_os = "macos")]
-    {
-        let home = dirs::home_dir().unwrap_or_default();
-        let plist_path = home.join("Library/LaunchAgents/ai.librefang.daemon.plist");
-        if plist_path.exists() {
-            let _ = std::process::Command::new("launchctl")
-                .args(["unload", &plist_path.to_string_lossy()])
-                .output();
-            match std::fs::remove_file(&plist_path) {
-                Ok(()) => ui::success("Removed LaunchAgent"),
-                Err(e) => ui::error(&format!("Failed to remove plist: {e}")),
-            }
-        } else {
-            ui::hint("No LaunchAgent found — nothing to remove.");
-        }
-    }
-    #[cfg(windows)]
-    {
-        let output = std::process::Command::new("reg")
-            .args([
-                "delete",
-                r"HKCU\Software\Microsoft\Windows\CurrentVersion\Run",
-                "/v",
-                "LibreFang",
-                "/f",
-            ])
-            .output();
-        match output {
-            Ok(o) if o.status.success() => {
-                ui::success("Removed from Windows startup");
-            }
-            _ => ui::hint("No startup entry found — nothing to remove."),
-        }
-    }
-    #[cfg(not(any(target_os = "linux", target_os = "macos", windows)))]
-    {
-        ui::error("Auto-start service is not supported on this platform.");
-    }
-}
-
-fn cmd_service_status() {
-    #[cfg(target_os = "linux")]
-    {
-        let home = dirs::home_dir().unwrap_or_default();
-        let service_path = home.join(".config/systemd/user/librefang.service");
-        if service_path.exists() {
-            ui::success("Systemd user service is registered");
-            // Show enabled/active status
-            if let Ok(output) = std::process::Command::new("systemctl")
-                .args(["--user", "is-enabled", "librefang.service"])
-                .output()
-            {
-                let status = String::from_utf8_lossy(&output.stdout).trim().to_string();
-                ui::kv("  Enabled", &status);
-            }
-            if let Ok(output) = std::process::Command::new("systemctl")
-                .args(["--user", "is-active", "librefang.service"])
-                .output()
-            {
-                let status = String::from_utf8_lossy(&output.stdout).trim().to_string();
-                ui::kv("  Active", &status);
-            }
-        } else {
-            ui::hint("No systemd user service registered.");
-            ui::hint("Run `librefang service install` to set it up.");
-        }
-    }
-    #[cfg(target_os = "macos")]
-    {
-        let home = dirs::home_dir().unwrap_or_default();
-        let plist_path = home.join("Library/LaunchAgents/ai.librefang.daemon.plist");
-        if plist_path.exists() {
-            ui::success("LaunchAgent is registered");
-            if let Ok(output) = std::process::Command::new("launchctl")
-                .args(["list"])
-                .output()
-            {
-                let stdout = String::from_utf8_lossy(&output.stdout);
-                let running = stdout.lines().any(|l| l.contains("ai.librefang.daemon"));
-                ui::kv("  Loaded", if running { "yes" } else { "not loaded" });
-            }
-        } else {
-            ui::hint("No LaunchAgent registered.");
-            ui::hint("Run `librefang service install` to set it up.");
-        }
-    }
-    #[cfg(windows)]
-    {
-        let output = std::process::Command::new("reg")
-            .args([
-                "query",
-                r"HKCU\Software\Microsoft\Windows\CurrentVersion\Run",
-                "/v",
-                "LibreFang",
-            ])
-            .output();
-        match output {
-            Ok(o) if o.status.success() => {
-                ui::success("Windows startup entry is registered");
-            }
-            _ => {
-                ui::hint("No startup entry registered.");
-                ui::hint("Run `librefang service install` to set it up.");
-            }
-        }
-    }
-    #[cfg(not(any(target_os = "linux", target_os = "macos", windows)))]
-    {
-        ui::error("Auto-start service is not supported on this platform.");
-    }
-}
-
-fn cmd_reset(confirm: bool) {
-    let librefang_dir = cli_librefang_home();
-
-    if !librefang_dir.exists() {
-        println!(
-            "Nothing to reset — {} does not exist.",
-            librefang_dir.display()
-        );
-        return;
-    }
-
-    if !confirm {
-        println!("  This will delete all data in {}", librefang_dir.display());
-        println!("  Including: config, database, agent manifests, credentials.");
-        println!();
-        let answer = prompt_input("  Are you sure? Type 'yes' to confirm: ");
-        if answer.trim() != "yes" {
-            println!("  Cancelled.");
-            return;
-        }
-    }
-
-    match std::fs::remove_dir_all(&librefang_dir) {
-        Ok(()) => ui::success(&i18n::t_args(
-            "reset-success",
-            &[("path", &librefang_dir.display().to_string())],
-        )),
-        Err(e) => {
-            ui::error(&i18n::t_args(
-                "reset-fail",
-                &[
-                    ("path", &librefang_dir.display().to_string()),
-                    ("error", &e.to_string()),
-                ],
-            ));
-            std::process::exit(1);
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Update
-// ---------------------------------------------------------------------------
-
-const RELEASE_REPO: &str = "librefang/librefang";
-const RELEASES_LATEST_API: &str =
-    "https://api.github.com/repos/librefang/librefang/releases/latest";
-const RELEASES_API: &str = "https://api.github.com/repos/librefang/librefang/releases";
-const SHELL_INSTALLER_URL: &str = "https://librefang.ai/install.sh";
-const POWERSHELL_INSTALLER_URL: &str = "https://librefang.ai/install.ps1";
-
-enum UpdateLaunch {
-    #[cfg(not(windows))]
-    Completed,
-    #[cfg(windows)]
-    Detached,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum ReleaseComparison {
-    Newer,
-    SameCore,
-    Older,
-    Unknown,
-}
-
-fn cmd_update(check: bool, version: Option<String>, channel_override: Option<String>) {
-    use librefang_types::config::UpdateChannel;
-
-    let current_exe = std::env::current_exe().unwrap_or_else(|e| {
-        ui::error(&format!("Cannot determine current executable path: {e}"));
-        std::process::exit(1);
-    });
-
-    let current_version = env!("CARGO_PKG_VERSION");
-    let current_exe_display = current_exe.display().to_string();
-    let requested_version = version.as_deref();
-
-    // Resolve update channel: CLI arg > config.toml > default (stable)
-    let channel = if let Some(ref ch) = channel_override {
-        match ch.parse::<UpdateChannel>() {
-            Ok(c) => c,
-            Err(e) => {
-                ui::error(&e);
-                std::process::exit(1);
-            }
-        }
-    } else {
-        load_update_channel_from_config().unwrap_or_default()
-    };
-
-    ui::section("Update");
-    ui::kv("Current", current_version);
-    ui::kv("Channel", &channel.to_string());
-    ui::kv("Binary", &current_exe_display);
-
-    let latest_tag = if requested_version.is_none() {
-        match fetch_latest_release_tag(channel) {
-            Ok(tag) => {
-                ui::kv("Latest", &tag);
-                Some(tag)
-            }
-            Err(err) => {
-                if check {
-                    ui::error(&format!("Failed to check latest release: {err}"));
-                    std::process::exit(1);
-                }
-                ui::warn_with_fix(
-                    &format!("Could not resolve the latest published release: {err}"),
-                    "Retry later, or pass `--version <tag>` to target a specific release.",
-                );
-                None
-            }
-        }
-    } else {
-        if let Some(target) = requested_version {
-            ui::kv("Target", target);
-        }
-        None
-    };
-    let target_tag = requested_version
-        .map(str::to_owned)
-        .or_else(|| latest_tag.clone());
-    let target_comparison = target_tag
-        .as_deref()
-        .map(|tag| compare_release_tag(tag, current_version));
-
-    if check {
-        match (target_tag.as_deref(), target_comparison) {
-            (Some(tag), Some(ReleaseComparison::Newer)) => {
-                ui::warn_with_fix(
-                    &format!("A newer published release is available: {tag}"),
-                    "Run `librefang update` to install it.",
-                );
-            }
-            (Some(tag), Some(ReleaseComparison::SameCore)) => {
-                ui::warn_with_fix(
-                    &format!(
-                        "The published release {tag} uses the same CLI version core as the current binary ({current_version})."
-                    ),
-                    "Run `librefang update` if you want the latest published build for this version line.",
-                );
-            }
-            (Some(tag), Some(ReleaseComparison::Older)) => {
-                ui::success(&format!(
-                    "Current binary version {current_version} is ahead of the published release {tag}."
-                ));
-            }
-            (Some(tag), Some(ReleaseComparison::Unknown)) => {
-                ui::warn_with_fix(
-                    &format!("Could not compare the current binary with release tag {tag}."),
-                    "If you want that exact release, run `librefang update --version <tag>`.",
-                );
-            }
-            _ => {
-                ui::warn_with_fix(
-                    "Unable to determine whether an update is available.",
-                    "Retry later when GitHub Releases is reachable.",
-                );
-            }
-        }
-        return;
-    }
-
-    if requested_version.is_none() {
-        match (latest_tag.as_deref(), target_comparison) {
-            (Some(tag), Some(ReleaseComparison::Older)) => {
-                ui::success(&format!(
-                    "Current binary version {current_version} is ahead of the latest published release {tag}."
-                ));
-                return;
-            }
-            (Some(tag), Some(ReleaseComparison::Unknown)) => {
-                ui::warn_with_fix(
-                    &format!(
-                        "Could not safely compare the current binary against release tag {tag}."
-                    ),
-                    &format!(
-                        "Re-run with `librefang update --version {tag}` to install it explicitly."
-                    ),
-                );
-                return;
-            }
-            _ => {}
-        }
-    }
-
-    let default_install = default_install_executable();
-    let cargo_install = cargo_install_executable();
-    let target_version = target_tag.as_deref();
-
-    #[cfg(windows)]
-    if same_path(&current_exe, &default_install) && find_daemon().is_some() {
-        ui::error_with_fix(
-            "Stop the running daemon before updating on Windows.",
-            "Run `librefang stop`, then `librefang update`, then `librefang start`.",
-        );
-        std::process::exit(1);
-    }
-
-    if same_path(&current_exe, &default_install) {
-        match run_official_update(target_version) {
-            #[cfg(not(windows))]
-            Ok(UpdateLaunch::Completed) => {
-                ui::success("LibreFang CLI updated.");
-                if let Some(installed) = installed_binary_version(&default_install) {
-                    ui::kv("Installed", &installed);
-                }
-                // Merge any new config defaults added in the updated binary.
-                // Spawn the new binary rather than calling cmd_init_upgrade() here,
-                // because the current process still holds the old binary's template.
-                ui::blank();
-                ui::hint("Merging new config defaults...");
-                let _ = std::process::Command::new(&default_install)
-                    .args(["init", "--upgrade"])
-                    .status();
-                ui::hint("If the daemon is running, restart it with `librefang restart`.");
-            }
-            #[cfg(windows)]
-            Ok(UpdateLaunch::Detached) => {
-                ui::success("Update launched in the background.");
-                ui::hint("Open a new terminal after it finishes and run `librefang --version`.");
-                ui::hint("If the daemon is running, restart it after the update completes.");
-            }
-            Err(err) => {
-                ui::error(&format!("Update failed: {err}"));
-                std::process::exit(1);
-            }
-        }
-        return;
-    }
-
-    if same_path(&current_exe, &cargo_install) {
-        let cargo_cmd = cargo_update_command(target_version);
-        ui::warn_with_fix(
-            "This binary was installed with cargo. Running `cargo install` from inside the active executable is intentionally blocked.",
-            &cargo_cmd,
-        );
-        return;
-    }
-
-    let official_path = default_install.display().to_string();
-    ui::warn_with_fix(
-        &format!(
-            "Automatic update only supports the official install path ({official_path}). This binary is running from a different location."
-        ),
-        &manual_installer_command(target_version),
-    );
-    ui::hint("If this binary came from another package manager, update it with that package manager instead.");
-}
-
-fn fetch_latest_release_tag(
-    channel: librefang_types::config::UpdateChannel,
-) -> Result<String, String> {
-    use librefang_types::config::UpdateChannel;
-
-    let client = update_http_client()?;
-
-    match channel {
-        UpdateChannel::Stable => {
-            // /releases/latest returns the latest non-draft, non-prerelease
-            let response = client
-                .get(RELEASES_LATEST_API)
-                .send()
-                .map_err(|e| format!("GitHub request failed: {e}"))?;
-            let status = response.status();
-            if !status.is_success() {
-                return Err(format!("GitHub API returned {status}"));
-            }
-            let body = response
-                .json::<serde_json::Value>()
-                .map_err(|e| format!("Failed to decode release metadata: {e}"))?;
-            body["tag_name"]
-                .as_str()
-                .filter(|tag| !tag.is_empty())
-                .map(str::to_string)
-                .ok_or_else(|| "Release metadata is missing `tag_name`".to_string())
-        }
-        UpdateChannel::Beta | UpdateChannel::Rc => {
-            // /releases lists all releases, newest first — filter by channel
-            let response = client
-                .get(RELEASES_API)
-                .send()
-                .map_err(|e| format!("GitHub request failed: {e}"))?;
-            let status = response.status();
-            if !status.is_success() {
-                return Err(format!("GitHub API returned {status}"));
-            }
-            let releases = response
-                .json::<Vec<serde_json::Value>>()
-                .map_err(|e| format!("Failed to decode releases list: {e}"))?;
-
-            for release in &releases {
-                let draft = release["draft"].as_bool().unwrap_or(false);
-                if draft {
-                    continue;
-                }
-                let Some(tag) = release["tag_name"].as_str().filter(|t| !t.is_empty()) else {
-                    continue;
-                };
-                match channel {
-                    UpdateChannel::Rc => return Ok(tag.to_string()),
-                    UpdateChannel::Beta => {
-                        if !tag.contains("-rc") {
-                            return Ok(tag.to_string());
-                        }
-                    }
-                    _ => unreachable!(),
-                }
-            }
-            Err(format!(
-                "No matching release found for the '{channel}' channel"
-            ))
-        }
-    }
-}
-
-fn update_http_client() -> Result<reqwest::blocking::Client, String> {
-    crate::http_client::client_builder()
-        .user_agent(format!("librefang-cli/{}", env!("CARGO_PKG_VERSION")))
-        .timeout(std::time::Duration::from_secs(60))
-        .build()
-        .map_err(|e| format!("Failed to build HTTP client: {e}"))
-}
-
-fn compare_release_tag(tag: &str, current_version: &str) -> ReleaseComparison {
-    let Some(release_core) = parse_version_core(normalize_release_tag(tag)) else {
-        return ReleaseComparison::Unknown;
-    };
-    let Some(current_core) = parse_version_core(current_version) else {
-        return ReleaseComparison::Unknown;
-    };
-
-    match release_core.cmp(&current_core) {
-        std::cmp::Ordering::Greater => ReleaseComparison::Newer,
-        std::cmp::Ordering::Equal => ReleaseComparison::SameCore,
-        std::cmp::Ordering::Less => ReleaseComparison::Older,
-    }
-}
-
-fn parse_version_core(version: &str) -> Option<Vec<u64>> {
-    let core = version.split('-').next()?;
-    if core.is_empty() {
-        return None;
-    }
-    core.split('.')
-        .map(|part| part.parse::<u64>().ok())
-        .collect()
-}
-
-fn run_official_update(version: Option<&str>) -> Result<UpdateLaunch, String> {
-    let script_url = if cfg!(windows) {
-        POWERSHELL_INSTALLER_URL
-    } else {
-        SHELL_INSTALLER_URL
-    };
-    let script = download_text(script_url)?;
-
-    #[cfg(windows)]
-    {
-        use std::os::windows::process::CommandExt;
-
-        const CREATE_NEW_PROCESS_GROUP: u32 = 0x0000_0200;
-        const DETACHED_PROCESS: u32 = 0x0000_0008;
-
-        let wrapped = format!(
-            "Start-Sleep -Seconds 1\r\n{script}\r\nRemove-Item $MyInvocation.MyCommand.Path -ErrorAction SilentlyContinue\r\n"
-        );
-        let script_path = write_update_script(&wrapped, "ps1")?;
-        let script_arg = script_path.to_string_lossy().to_string();
-
-        let mut command = std::process::Command::new("powershell");
-        command
-            .args([
-                "-NoProfile",
-                "-ExecutionPolicy",
-                "Bypass",
-                "-File",
-                &script_arg,
-            ])
-            .stdin(std::process::Stdio::null())
-            .stdout(std::process::Stdio::null())
-            .stderr(std::process::Stdio::null())
-            .creation_flags(CREATE_NEW_PROCESS_GROUP | DETACHED_PROCESS);
-        if let Some(tag) = version {
-            command.env("LIBREFANG_VERSION", tag);
-        }
-
-        command
-            .spawn()
-            .map_err(|e| format!("Failed to launch PowerShell updater: {e}"))?;
-        Ok(UpdateLaunch::Detached)
-    }
-
-    #[cfg(not(windows))]
-    {
-        let script_path = write_update_script(&script, "sh")?;
-        let mut command = std::process::Command::new("sh");
-        command.arg(&script_path);
-        if let Some(tag) = version {
-            command.env("LIBREFANG_VERSION", tag);
-        }
-
-        let status = command
-            .status()
-            .map_err(|e| format!("Failed to run installer: {e}"))?;
-        let _ = std::fs::remove_file(&script_path);
-        if !status.success() {
-            return Err(format!("Installer exited with status {status}"));
-        }
-        Ok(UpdateLaunch::Completed)
-    }
-}
-
-fn download_text(url: &str) -> Result<String, String> {
-    let client = update_http_client()?;
-    let response = client
-        .get(url)
-        .send()
-        .map_err(|e| format!("Download failed: {e}"))?;
-    let status = response.status();
-    if !status.is_success() {
-        return Err(format!("Download returned {status}"));
-    }
-    response
-        .text()
-        .map_err(|e| format!("Failed to read response body: {e}"))
-}
-
-#[cfg(not(windows))]
-fn installed_binary_version(path: &std::path::Path) -> Option<String> {
-    let output = std::process::Command::new(path)
-        .arg("--version")
-        .output()
-        .ok()?;
-    if !output.status.success() {
-        return None;
-    }
-
-    let version = String::from_utf8_lossy(&output.stdout).trim().to_string();
-    if version.is_empty() {
-        None
-    } else {
-        Some(version)
-    }
-}
-
-fn write_update_script(contents: &str, extension: &str) -> Result<PathBuf, String> {
-    let unique = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_millis())
-        .unwrap_or(0);
-    let path = std::env::temp_dir().join(format!(
-        "librefang-update-{}-{unique}.{extension}",
-        std::process::id()
-    ));
-    std::fs::write(&path, contents).map_err(|e| format!("Failed to write updater script: {e}"))?;
-    restrict_file_permissions(&path);
-    Ok(path)
-}
-
-fn default_install_executable() -> PathBuf {
-    cli_librefang_home().join("bin").join(binary_name())
-}
-
-fn cargo_install_executable() -> PathBuf {
-    dirs::home_dir()
-        .unwrap_or_else(std::env::temp_dir)
-        .join(".cargo")
-        .join("bin")
-        .join(binary_name())
-}
-
-fn binary_name() -> &'static str {
-    if cfg!(windows) {
-        "librefang.exe"
-    } else {
-        "librefang"
-    }
-}
-
-fn same_path(left: &std::path::Path, right: &std::path::Path) -> bool {
-    let left = std::fs::canonicalize(left).unwrap_or_else(|_| left.to_path_buf());
-    let right = std::fs::canonicalize(right).unwrap_or_else(|_| right.to_path_buf());
-    left == right
-}
-
-fn normalize_release_tag(tag: &str) -> &str {
-    tag.strip_prefix('v').unwrap_or(tag)
-}
-
-fn cargo_update_command(version: Option<&str>) -> String {
-    match version {
-        Some(tag) => format!(
-            "cargo install --git https://github.com/{RELEASE_REPO} --tag {tag} librefang-cli --force"
-        ),
-        None => format!(
-            "cargo install --git https://github.com/{RELEASE_REPO} librefang-cli --force"
-        ),
-    }
-}
-
-fn manual_installer_command(version: Option<&str>) -> String {
-    #[cfg(windows)]
-    {
-        match version {
-            Some(tag) => {
-                format!("$env:LIBREFANG_VERSION='{tag}'; irm {POWERSHELL_INSTALLER_URL} | iex")
-            }
-            None => format!("irm {POWERSHELL_INSTALLER_URL} | iex"),
-        }
-    }
-
-    #[cfg(not(windows))]
-    {
-        match version {
-            Some(tag) => format!("curl -fsSL {SHELL_INSTALLER_URL} | LIBREFANG_VERSION={tag} sh"),
-            None => format!("curl -fsSL {SHELL_INSTALLER_URL} | sh"),
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Uninstall
-// ---------------------------------------------------------------------------
-
-fn cmd_uninstall(confirm: bool, keep_config: bool) {
-    let librefang_dir = cli_librefang_home();
-    let exe_path = std::env::current_exe().ok();
-
-    // Step 1: Show what will be removed
-    println!();
-    println!(
-        "  {}",
-        "This will completely uninstall LibreFang from your system."
-            .bold()
-            .red()
-    );
-    println!();
-    if librefang_dir.exists() {
-        if keep_config {
-            println!(
-                "  • Remove data in {} (keeping config files)",
-                librefang_dir.display()
-            );
-        } else {
-            println!("  • Remove {}", librefang_dir.display());
-        }
-    }
-    if let Some(ref exe) = exe_path {
-        println!("  • Remove binary: {}", exe.display());
-    }
-    // Check cargo bin path
-    let cargo_bin = dirs::home_dir()
-        .unwrap_or_else(std::env::temp_dir)
-        .join(".cargo")
-        .join("bin")
-        .join(if cfg!(windows) {
-            "librefang.exe"
-        } else {
-            "librefang"
-        });
-    if cargo_bin.exists() && exe_path.as_ref().is_none_or(|e| *e != cargo_bin) {
-        println!("  • Remove cargo binary: {}", cargo_bin.display());
-    }
-    println!("  • Remove auto-start entries (if any)");
-    println!("  • Clean PATH from shell configs (if any)");
-    println!();
-
-    // Step 2: Confirm
-    if !confirm {
-        let answer = prompt_input("  Type 'uninstall' to confirm: ");
-        if answer.trim() != "uninstall" {
-            println!("  Cancelled.");
-            return;
-        }
-        println!();
-    }
-
-    // Step 3: Stop running daemon
-    if find_daemon().is_some() {
-        println!("  {}", i18n::t("uninstall-stopping-daemon"));
-        cmd_stop(None);
-        // Give it a moment
-        std::thread::sleep(std::time::Duration::from_secs(1));
-        // Force kill if still alive
-        if find_daemon().is_some() {
-            if let Some(info) = read_daemon_info(&librefang_dir) {
-                force_kill_pid(info.pid);
-                let _ = std::fs::remove_file(librefang_dir.join("daemon.json"));
-            }
-        }
-    }
-
-    // Step 4: Remove auto-start entries
-    let user_home = dirs::home_dir().unwrap_or_else(std::env::temp_dir);
-    remove_autostart_entries(&user_home);
-
-    // Step 5: Clean PATH from shell configs
-    if let Some(ref exe) = exe_path {
-        if let Some(bin_dir) = exe.parent() {
-            clean_path_entries(&user_home, &bin_dir.to_string_lossy());
-        }
-    }
-
-    // Step 6: Remove ~/.librefang/ data
-    if librefang_dir.exists() {
-        if keep_config {
-            remove_dir_except_config(&librefang_dir);
-            ui::success(&i18n::t("uninstall-removed-data-kept"));
-        } else {
-            match std::fs::remove_dir_all(&librefang_dir) {
-                Ok(()) => ui::success(&i18n::t_args(
-                    "uninstall-removed",
-                    &[("path", &librefang_dir.display().to_string())],
-                )),
-                Err(e) => ui::error(&i18n::t_args(
-                    "uninstall-remove-failed",
-                    &[
-                        ("path", &librefang_dir.display().to_string()),
-                        ("error", &e.to_string()),
-                    ],
-                )),
-            }
-        }
-    }
-
-    // Step 7: Remove cargo bin copy if it exists and is separate from current exe
-    if cargo_bin.exists() && exe_path.as_ref().is_none_or(|e| *e != cargo_bin) {
-        match std::fs::remove_file(&cargo_bin) {
-            Ok(()) => ui::success(&i18n::t_args(
-                "uninstall-removed",
-                &[("path", &cargo_bin.display().to_string())],
-            )),
-            Err(e) => ui::error(&i18n::t_args(
-                "uninstall-remove-failed",
-                &[
-                    ("path", &cargo_bin.display().to_string()),
-                    ("error", &e.to_string()),
-                ],
-            )),
-        }
-    }
-
-    // Step 8: Remove the binary itself (skip if already removed with ~/.librefang/)
-    if let Some(exe) = exe_path {
-        if exe.exists() {
-            remove_self_binary(&exe);
-        }
-    }
-
-    println!();
-    ui::success(&i18n::t("uninstall-goodbye"));
-}
-
-/// Remove auto-start / launch-agent / systemd entries.
-#[allow(unused_variables)]
-fn remove_autostart_entries(home: &std::path::Path) {
-    #[cfg(windows)]
-    {
-        // Windows: remove from HKCU\Software\Microsoft\Windows\CurrentVersion\Run
-        let output = std::process::Command::new("reg")
-            .args([
-                "delete",
-                r"HKCU\Software\Microsoft\Windows\CurrentVersion\Run",
-                "/v",
-                "LibreFang",
-                "/f",
-            ])
-            .output();
-        match output {
-            Ok(o) if o.status.success() => {
-                ui::success(&i18n::t("uninstall-removed-autostart-win"));
-            }
-            _ => {} // Entry didn't exist — that's fine
-        }
-    }
-
-    #[cfg(target_os = "macos")]
-    {
-        let plist = home.join("Library/LaunchAgents/ai.librefang.desktop.plist");
-        if plist.exists() {
-            // Unload first
-            let _ = std::process::Command::new("launchctl")
-                .args(["unload", &plist.to_string_lossy()])
-                .output();
-            match std::fs::remove_file(&plist) {
-                Ok(()) => ui::success(&i18n::t("uninstall-removed-launch-agent")),
-                Err(e) => ui::error(&i18n::t_args(
-                    "uninstall-remove-launch-fail",
-                    &[("error", &e.to_string())],
-                )),
-            }
-        }
-    }
-
-    #[cfg(target_os = "linux")]
-    {
-        let desktop_file = home.join(".config/autostart/LibreFang.desktop");
-        if desktop_file.exists() {
-            match std::fs::remove_file(&desktop_file) {
-                Ok(()) => ui::success(&i18n::t("uninstall-removed-autostart-linux")),
-                Err(e) => ui::error(&i18n::t_args(
-                    "uninstall-remove-autostart-fail",
-                    &[("error", &e.to_string())],
-                )),
-            }
-        }
-
-        // Also check for systemd user service
-        let service_file = home.join(".config/systemd/user/librefang.service");
-        if service_file.exists() {
-            let _ = std::process::Command::new("systemctl")
-                .args(["--user", "disable", "--now", "librefang.service"])
-                .output();
-            match std::fs::remove_file(&service_file) {
-                Ok(()) => {
-                    let _ = std::process::Command::new("systemctl")
-                        .args(["--user", "daemon-reload"])
-                        .output();
-                    ui::success(&i18n::t("uninstall-removed-systemd"));
-                }
-                Err(e) => ui::error(&i18n::t_args(
-                    "uninstall-remove-systemd-fail",
-                    &[("error", &e.to_string())],
-                )),
-            }
-        }
-    }
-}
-
-/// Remove lines from shell config files that add librefang to PATH.
-#[allow(unused_variables)]
-fn clean_path_entries(home: &std::path::Path, librefang_dir: &str) {
-    #[cfg(not(windows))]
-    {
-        let shell_files = [
-            home.join(".bashrc"),
-            home.join(".bash_profile"),
-            home.join(".profile"),
-            home.join(".zshrc"),
-            home.join(".config/fish/config.fish"),
-        ];
-
-        for path in &shell_files {
-            if !path.exists() {
-                continue;
-            }
-            let Ok(content) = std::fs::read_to_string(path) else {
-                continue;
-            };
-            let filtered: Vec<&str> = content
-                .lines()
-                .filter(|line| !is_librefang_path_line(line, librefang_dir))
-                .collect();
-            if filtered.len() < content.lines().count() {
-                let new_content = filtered.join("\n");
-                // Preserve trailing newline if original had one
-                let new_content = if content.ends_with('\n') {
-                    format!("{new_content}\n")
-                } else {
-                    new_content
-                };
-                if std::fs::write(path, &new_content).is_ok() {
-                    ui::success(&i18n::t_args(
-                        "uninstall-cleaned-path",
-                        &[("path", &path.display().to_string())],
-                    ));
-                }
-            }
-        }
-    }
-
-    #[cfg(windows)]
-    {
-        // Read User PATH via PowerShell, filter out librefang entries, write back
-        let output = std::process::Command::new("powershell")
-            .args([
-                "-NoProfile",
-                "-Command",
-                "[Environment]::GetEnvironmentVariable('PATH', 'User')",
-            ])
-            .output();
-        if let Ok(out) = output {
-            if out.status.success() {
-                let current = String::from_utf8_lossy(&out.stdout);
-                let current = current.trim();
-                if !current.is_empty() {
-                    let dir_lower = librefang_dir.to_lowercase();
-                    let filtered: Vec<&str> = current
-                        .split(';')
-                        .filter(|entry| {
-                            let e = entry.trim().to_lowercase();
-                            !e.is_empty() && !e.contains("librefang") && !e.contains(&dir_lower)
-                        })
-                        .collect();
-                    if filtered.len() < current.split(';').count() {
-                        let new_path = filtered.join(";");
-                        let ps_cmd = format!(
-                            "[Environment]::SetEnvironmentVariable('PATH', '{}', 'User')",
-                            new_path.replace('\'', "''")
-                        );
-                        let result = std::process::Command::new("powershell")
-                            .args(["-NoProfile", "-Command", &ps_cmd])
-                            .output();
-                        if result.is_ok_and(|o| o.status.success()) {
-                            ui::success(&i18n::t("uninstall-cleaned-path-win"));
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Returns true if a shell config line is an librefang PATH export.
-/// Must match BOTH an librefang reference AND a PATH-setting pattern.
-#[cfg(any(not(windows), test))]
-fn is_librefang_path_line(line: &str, librefang_dir: &str) -> bool {
-    let lower = line.to_lowercase();
-    let has_librefang =
-        lower.contains("librefang") || lower.contains(&librefang_dir.to_lowercase());
-    if !has_librefang {
-        return false;
-    }
-    // Match common PATH-setting patterns
-    lower.contains("export path=")
-        || lower.contains("export path =")
-        || lower.starts_with("path=")
-        || lower.contains("set -gx path")
-        || lower.contains("fish_add_path")
-}
-
-/// Remove everything in ~/.librefang/ except config files.
-fn remove_dir_except_config(librefang_dir: &std::path::Path) {
-    let keep = ["config.toml", ".env", "secrets.env"];
-    let Ok(entries) = std::fs::read_dir(librefang_dir) else {
-        return;
-    };
-    for entry in entries.flatten() {
-        let name = entry.file_name();
-        let name_str = name.to_string_lossy();
-        if keep.contains(&name_str.as_ref()) {
-            continue;
-        }
-        let path = entry.path();
-        if path.is_dir() {
-            let _ = std::fs::remove_dir_all(&path);
-        } else {
-            let _ = std::fs::remove_file(&path);
-        }
-    }
-}
-
-/// Remove the currently-running binary.
-fn remove_self_binary(exe_path: &std::path::Path) {
-    #[cfg(unix)]
-    {
-        // On Unix, running binaries can be unlinked — the OS keeps the inode
-        // alive until the process exits.
-        match std::fs::remove_file(exe_path) {
-            Ok(()) => ui::success(&i18n::t_args(
-                "uninstall-removed",
-                &[("path", &exe_path.display().to_string())],
-            )),
-            Err(e) => ui::error(&i18n::t_args(
-                "uninstall-remove-failed",
-                &[
-                    ("path", &exe_path.display().to_string()),
-                    ("error", &e.to_string()),
-                ],
-            )),
-        }
-    }
-
-    #[cfg(windows)]
-    {
-        // Windows locks running executables. Rename first, then spawn a
-        // detached process that waits briefly and deletes the renamed file.
-        let old_path = exe_path.with_extension("exe.old");
-        if std::fs::rename(exe_path, &old_path).is_err() {
-            ui::error(&format!(
-                "Could not rename binary for deferred deletion: {}",
-                exe_path.display()
-            ));
-            return;
-        }
-
-        use std::os::windows::process::CommandExt;
-        const CREATE_NEW_PROCESS_GROUP: u32 = 0x0000_0200;
-        const DETACHED_PROCESS: u32 = 0x0000_0008;
-
-        let del_cmd = format!(
-            "ping -n 3 127.0.0.1 >nul & del /f /q \"{}\"",
-            old_path.display()
-        );
-        let _ = std::process::Command::new("cmd.exe")
-            .args(["/C", &del_cmd])
-            .creation_flags(CREATE_NEW_PROCESS_GROUP | DETACHED_PROCESS)
-            .spawn();
-
-        ui::success(&i18n::t_args(
-            "uninstall-removed",
-            &[("path", &exe_path.display().to_string())],
-        ));
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::{
